% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-conversion-functions.R
\name{as_adj_matrix}
\alias{as_adj_matrix}
\alias{as_adj_matrix.parentslist}
\title{Convert to an adjacency matrix}
\usage{
as_adj_matrix(x)

\method{as_adj_matrix}{parentslist}(x)
}
\arguments{
\item{x}{an R object}
}
\value{
the equivalent adjacency matrix
}
\description{
Convert to an adjacency matrix
}
