% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_bn.R
\name{as_bn}
\alias{as_bn}
\alias{as_bn.parentslist}
\alias{as_bn.sevt}
\title{Convert to a \pkg{bnlearn} object}
\usage{
as_bn(x)

\method{as_bn}{parentslist}(x)

\method{as_bn}{sevt}(x)
}
\arguments{
\item{x}{an R object of class \code{sevt} or \code{parentslist}.}
}
\value{
an object of class \code{bn} from package \pkg{bnlearn}.
}
\description{
Convert a staged tree object into an object of class \code{bn}
from the \pkg{bnlearn} package.
}
