% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{as_parentslist}
\alias{as_parentslist}
\title{Obtain the equivalent DAG list of parents}
\usage{
as_parentslist(x)
}
\arguments{
\item{x}{an object of class \code{sevt}}
}
\value{
A list where the parents of each variable are
listed, NAs are reported if a variable has no parents.
}
\description{
Compute, for each variable in the staged tree,
the parent set in the equivalent DAG.
}
