% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{as_sevt.bn.fit}
\alias{as_sevt.bn.fit}
\title{Convert a bn fit to a staged event tree object}
\usage{
\method{as_sevt}{bn.fit}(x, order = NULL)
}
\arguments{
\item{x}{a \code{bn.fit} object.}

\item{order}{a topological order of the variables,
WARNING: the order will not be checked.}
}
\value{
A staged event tree object.
}
\description{
Convert a bn fit to a staged event tree object
}
