% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{Asym}
\alias{Asym}
\title{Asym dataset}
\format{
A data frame with 1000 observations of 4 binary variables.
}
\source{
The data has been manually generated by Federico Carli \email{carli@dima.unige}.
}
\usage{
Asym
}
\description{
Artificial dataset with observations from four variables
having a non-symmetrical conditional independence structure.
}
\keyword{datasets}
