% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-model-selection.R
\name{backward_hill_climb_random}
\alias{backward_hill_climb_random}
\title{Backword random hill-climbing}
\usage{
backward_hill_climb_random(object = NULL, score = function(x)
  return(-BIC(x)), eps = 1e-04, max_iter = 100, trace = 0)
}
\arguments{
\item{object}{a staged event tree model}

\item{score}{the score function to be maximized}

\item{eps}{the stopping criteria for the relative score increase}

\item{max_iter}{the maximum number of iteration}

\item{trace}{if >0 increasingly amount of info 
is printed (via \code{message})}
}
\description{
Randomly try to join stages
}
