% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-model-selection.R
\name{backward_joining_KL}
\alias{backward_joining_KL}
\title{Backword joining of stages}
\usage{
backward_joining_KL(object = NULL, thr = 0.01, verbose = FALSE)
}
\arguments{
\item{object}{the staged event tree from where to start}

\item{thr}{the threshold for joining stages}

\item{verbose}{logical}
}
\value{
the learned staged event tree
}
\description{
Join stages from more complex to simpler models
using entropy do decide if join or not a stage
}
