% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-plot-functions.R
\name{barplot.sevt}
\alias{barplot.sevt}
\title{Barplot method for staged event trees}
\usage{
\method{barplot}{sevt}(object, var, legend.text = FALSE, ...)
}
\arguments{
\item{object}{a staged tree object}

\item{var}{name of varaibles in the model}
}
\description{
Barplot method for staged event trees
}
