% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-plot-functions.R
\name{barplot_stages}
\alias{barplot_stages}
\title{Barplot of stage probabilities}
\usage{
barplot_stages(object, var, legend.text = FALSE, col = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{sevt}.}

\item{var}{name of a variable in \code{object}.}

\item{legend.text}{logical.}

\item{col}{color mapping for the stages, see \code{col}
argument in \code{\link{plot.sevt}}.}

\item{...}{additional arguments passed to \code{\link{barplot}}.}
}
\description{
Barplot of stage probabilities
}
\examples{
model <- fbhc(full(PhDArticles, lambda = 1))
barplot_stages(model, "Kids", beside = TRUE)
}
