% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-model-selection.R
\name{bhc_sevt}
\alias{bhc_sevt}
\title{Backward hill-climbing}
\usage{
bhc_sevt(
  object,
  score = function(x) {     return(-BIC(x)) },
  max_iter = Inf,
  scope = NULL,
  ignore = NULL,
  trace = 0
)
}
\arguments{
\item{object}{a staged event tree model}

\item{score}{the score function to be maximized}

\item{max_iter}{the maximum number of iterations per variable}

\item{scope}{names of variables that should be considered for the optimization}

\item{ignore}{stage names that should be ignored}

\item{trace}{if >0 increasingly amount of info
is printed (via \code{message})}
}
\value{
The final staged event tree obtained.
}
\description{
Backward Hill-climbing search of staged event trees with
iterative joining of stages
}
\details{
For each variable the algorithm try to join stages
and move to the best model that increase the score. When no
increase is possible it moves to the next variable.
}
\examples{
DD <- generate_xor_dataset(n = 4, N = 100)
model <- bhc_sevt(full(DD), trace = 2)
summary(model)
}
