% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-model-selection.R
\name{bj.sevt}
\alias{bj.sevt}
\title{Backword joining of stages}
\usage{
bj.sevt(object = NULL, distance = kl, thr = 0.1, trace = 0, ...)
}
\arguments{
\item{object}{the staged event tree from where to start}

\item{distance}{the distance between probabilities to use}

\item{thr}{the threshold for joining stages}

\item{trace}{if >0 increasingly amount of info}

\item{...}{additional parameters to be passed to the distance function
is printed (via \code{message})}
}
\value{
the learned staged event tree
}
\description{
Join stages from more complex to simpler models
using a distance and a threshold value
}
\examples{

########
DD <- generate_xor_dataset(n = 5, N = 1000)
model_full <- staged_ev_tree(DD, fit = TRUE, full = TRUE, lambda = 1)
model <- bj.sevt(model_full, trace = 2)
}
