% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-ceg-functions.R
\name{ceg}
\alias{ceg}
\title{Chain event graph (CEG)}
\usage{
ceg(object)
}
\arguments{
\item{object}{an object of class \code{sevt}.}
}
\value{
an object of class \code{ceg}.
}
\description{
Build the CEG representation from an object of class \code{\link{sevt}}.
}
\details{
An object of class \code{ceg} is a staged event tree object with
additional information on the positions.
}
\examples{
DD <- generate_xor_dataset(3, 100)
model <- stages_bhc(full(DD))
model.ceg <- ceg(model)
model.ceg$positions
}
