% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-ceg-functions.R
\name{ceg.sevt}
\alias{ceg.sevt}
\title{Creates the CEG starting from a staged tree}
\usage{
ceg.sevt(object)
}
\arguments{
\item{object}{a staged event tree}
}
\value{
a ceg object
}
\description{
Creates the CEG starting from a staged tree
}
\details{
A ceg object is a staged event tree object with
additional information on the
positions.
}
