% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-ceg-functions.R
\name{ceg2adjmat}
\alias{ceg2adjmat}
\title{Ceg to adjmat of graph}
\usage{
ceg2adjmat(x, ignore = x$name_unobserved)
}
\arguments{
\item{x}{an object of class \code{\link{ceg}}.}

\item{ignore}{vector of stages which will be ignored and left untouched,
by default the name of the unobserved stages stored in
\code{x$name_unobserved}.}
}
\value{
the adj matrix
}
\description{
Obtain the adjacency matrix corresponding to a CEG.
}
\details{
This utility function can be used to prepare the adjacency
matrix to plot the CEG using a graph package (e.g. \pkg{igraph}).
}
\examples{
model <- stages_fbhc(full(PhDArticles))
model.ceg <- ceg(model)
ceg2adjmat(model.ceg)
}
