% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{CHDS}
\alias{CHDS}
\title{Child Health and Development Studies}
\format{A data frame with 890 rows and 4 variables:
\describe{
  \item{Social}{Binary variable}
  \item{Economic}{Binary variable}
  \item{Events}{Ternary variable}
  \item{Admission}{Binary variable}
}}
\source{
\url{FIND URL}
}
\usage{
CHDS
}
\description{
A dataset containing ... (to do)
}
\keyword{datasets}
