% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-base-model-function.R
\name{check_sevt}
\alias{check_sevt}
\alias{check_sevt_ctables}
\alias{check_sevt_fit}
\title{check sevt object}
\usage{
check_sevt(object)

check_sevt_ctables(object)

check_sevt_fit(object)
}
\arguments{
\item{object}{an object of class sevt}
}
\description{
check sevt object
}
\keyword{internal}
