% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-classifier-model-functions.R
\name{classifier_staged_ev_tree}
\alias{classifier_staged_ev_tree}
\title{Classifer model with staged event tree}
\usage{
classifier_staged_ev_tree(x, ...)
}
\description{
Classifer model with staged event tree
}
