% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{compare.sevt}
\alias{compare.sevt}
\title{Compare two staged event tree}
\usage{
compare.sevt(object1, object2, tree = FALSE, plot = FALSE, ...)
}
\arguments{
\item{object1}{a staged event tree}

\item{object2}{a staged event tree}

\item{tree}{logical, if \code{TRUE} the difference tree is returned}

\item{plot}{logical}

\item{...}{additional paraters to be passed to \code{\link{plot.sevt}}}
}
\value{
Logical or a difference tree (if \code{tree = TRUE})
}
\description{
Compare two staged event tree
}
\examples{

#########
data("Trump")
mod1 <- bhc.sevt(full(Trump, lambda = 1))
mod2 <- fbhc.sevt(full(Trump, lambda = 1))
compare.sevt(mod1, mod2)
}
