% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint1}
\alias{confint1}
\title{Confidence intervals for staged event tree parameters}
\usage{
confint1(
  object,
  level = 0.95,
  sides = c("two.sided", "left", "right"),
  method = c("goodman", "wald", "waldcc", "wilson"),
  lambda = object$lambda,
  digits = 4,
  var = "all",
  stage = "all",
  ...
)
}
\arguments{
\item{object}{an object of class \code{sevt}.}

\item{level}{the confidence level desired.}

\item{sides}{a character string specifying the side of the confidence interval, must be one of "two.sided"
(default), "left" or "right".}

\item{method}{a character string specifing which method to use: "goodman", "wald", "waldcc" or "wilson".}

\item{lambda}{smoothing coefficient (default object$lambda).}

\item{digits}{number of digits for the confidence intervals; default is 4.}

\item{var}{name of the variables in \code{object} for which the confidence intervals has to be calculated.
Default is "all", which computes confidence intervals for all the variables in \code{object}.}

\item{stage}{character vector, the name of the stages for \code{var} for which the confidence intervals
should be calculated. Default is "all", which computes confidence intervals
for all the stages for variable \code{var} in \code{object}.}

\item{...}{arguments for compatibility.}
}
\value{
An object of class \code{confint.sevt} containing the desired confidence intervals.
}
\description{
Confint method for class \code{sevt}.
}
\details{
Compute confidence intervals for staged event trees. By default, it computes two-sided confidence intervals
at 95\% level for the whole tree given in \code{object} with the wald method. If a single variable is given
in \code{var}, \code{stage} can be specified as a character vector of stages, while if multiple variables are
given in \code{var}, \code{stage} should be only set to \code{"all"} (this is done automatically in the function).
}
\examples{
m1 <- stages_bj(full(PhDArticles), distance = "kullback", thr = 0.01)
confint1(m1)
confint1(m1, level = 0.90, sides = "left", var = "Prestige")
confint1(m1, var = "Married", stage = c(1,2,6))
confint1(m1, var = c("Married", "Kids"), stage = c(9:12))
}
\references{
Goodman, L. A. (1965) On Simultaneous Confidence Intervals for Multinomial Proportions Technometrics, 7, 247-254.

Wald, A. Tests of statistical hypotheses concerning several parameters when the number of observations is large, Trans. Am. Math. Soc. 54 (1943) 426-482.

Wilson, E. B. Probable inference, the law of succession and statistical inference, J.Am. Stat. Assoc. 22 (1927) 209-212.
}
