% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_sevt.R
\name{copy_sevt}
\alias{copy_sevt}
\title{Copy a sevt obejct}
\usage{
copy_sevt(object, extra = NULL)
}
\arguments{
\item{object}{an object of class \code{sevt}.}

\item{extra}{additional arguments.}
}
\value{
the copied object.
}
\description{
Copy a sevt obejct
}
