% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-util-functions.R
\name{cs_matrices}
\alias{cs_matrices}
\title{Context-specific stages matrices}
\usage{
cs_matrices(object, v)
}
\arguments{
\item{object}{an object of class sevt}

\item{v}{string, the name of one of the variables in \code{object}}
}
\value{
A list with \code{i-1} matrices, where \code{i} is the depth
of variable \code{v} in the tree.
}
\description{
Generate the sequence of all the context-specific
stages matrices for a given variable in the model.
}
\keyword{internal}
