% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{df_sevt}
\alias{df_sevt}
\title{Number of parameters of a staged event tree}
\usage{
df_sevt(x)
}
\arguments{
\item{x}{An object of class \code{sevt}.}
}
\value{
integer, degrees of freedom of the staged event tree.
}
\description{
Return the number of parameters of the model.
}
\examples{

#########
mod_f <- full(PhDArticles)
df_sevt(mod_f)

mod_i <- indep(PhDArticles)
df_sevt(mod_i)
}
