% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-plot-functions.R
\name{edge}
\alias{edge}
\title{Plot an edge}
\usage{
edge(from, to, label = "", col = "black", ...)
}
\arguments{
\item{from}{From}

\item{to}{To}

\item{label}{the label}

\item{col}{color}

\item{...}{additional parameters passed to \code{par()}}
}
\description{
Plot an edge
}
