% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{effindep.sevt}
\alias{effindep.sevt}
\title{Built (efficently) an independent event tree}
\usage{
effindep.sevt(x, lambda, ctables = TRUE)
}
\arguments{
\item{x}{data}

\item{lambda}{the smoothing parameter}

\item{ctables}{logical, if \code{TRUE} ctables will be computed and attached}
}
\value{
A fitted staged event tree with one stage per variable
}
\description{
Built (efficently) an independent event tree
}
\examples{
DD <- generate_xor_dataset(15, 100)
system.time(model1 <- staged_ev_tree(DD, fit = TRUE, lambda = 1))
system.time(model2 <- effindep.sevt(DD, lambda = 1, ctables = TRUE))
model1
model2
}
