% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-util-functions.R
\name{entr}
\alias{entr}
\title{compute the entropy}
\usage{
entr(p)
}
\arguments{
\item{p}{the vector of normalized probability}
}
\value{
the entropy as \code{-sum(p*log(p))}
}
\description{
compute the entropy
}
