% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erase_fit.R
\name{erase_fit}
\alias{erase_fit}
\title{Erase the sevt fit}
\usage{
erase_fit(object)
}
\arguments{
\item{object}{an object of class \code{sevt}}
}
\value{
an object of class \code{sevt} without
\code{prob} and \code{ll} field.
}
\description{
Erase the sevt fit
}
\keyword{internal}
