% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-custom-functions.R
\name{exhaustive_ordered_search}
\alias{exhaustive_ordered_search}
\title{Full search of an ordered variable.}
\usage{
exhaustive_ordered_search(
  object,
  n_bins = 3,
  variable = NULL,
  ignore = object$name_unobserved,
  score = function(x) {     return(logLik(x)) }
)
}
\arguments{
\item{object}{an object of class \code{sevt} with fitted probabilities and
data, as returned by \code{full} or \code{sevt_fit}.}

\item{n_bins}{final number of stages.}

\item{variable}{name of a variables that should be considered for the optimization.}

\item{ignore}{vector of stages which will be ignored and left untouched,
by default the name of the unobserved stages stored in
\code{object$name_unobserved}.}

\item{score}{the score function to be maximized.}
}
\value{
The final staged event tree obtained.
}
\description{
Full search on one level of a staged event tree with
iterative joining of "adjacent" stage for an ordered variable.
}
\details{
For the given variable the algorithm separates the stages into bins
and joins all stages in one bin. It searches all possible splittings of the
stages into the bins and returns the best scoring model.
}
