% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-base-model-function.R
\name{expand_prob}
\alias{expand_prob}
\title{Expand probabilities of a staged event tree}
\usage{
expand_prob(object)
}
\arguments{
\item{object}{a fitted staged event tree object}
}
\value{

}
\description{
Expand probabilities of a staged event tree
}
\details{
This function build a stratified event tree object
from a staged event tree.
}
\keyword{internal}
