% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-model-selection.R
\name{fast_backward_hill_climb}
\alias{fast_backward_hill_climb}
\title{Fast backword hill-climbing}
\usage{
fast_backward_hill_climb(object = NULL, score = function(x)
  return(-BIC(x)), max_iter = Inf, verbose = FALSE)
}
\arguments{
\item{object}{a staged event tree model}

\item{score}{the score function to be maximized}

\item{max_iter}{the maximum number of iteration}

\item{verbose}{If info should be printed (via \code{message})}
}
\description{
Move to the first model that increase the score
}
