% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parentslist.R
\name{find_roots}
\alias{find_roots}
\title{find roots in graphs}
\usage{
find_roots(x)
}
\arguments{
\item{x}{a graph encoded as list of parents.}
}
\value{
nodes with 0 incoming edges, that is with 0 in-degree.
}
\description{
find roots in graphs
}
