% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{fit.staged_ev_tree}
\alias{fit.staged_ev_tree}
\title{Fit a staged event tree}
\usage{
fit.staged_ev_tree(sevt, data = NULL, lambda = 0, ...)
}
\arguments{
\item{sevt}{The staged event tree object to be fitted}

\item{data}{the data.frame used to fit the staged event tree}

\item{lambda}{the laplace smoothing}

\item{...}{additional parameters}
}
\value{
A staged event tree object with the conditional probabilities fitted
}
\description{
Fit a staged event tree
}
\examples{
model <- staged_ev_tree(list(X = c("good", "bad"), Y = c("high", "low")))
D <- data.frame(X = c("good", "good", "bad"),
                Y = c("high", "low", "low"))
model.fit <- fit.staged_ev_tree(model, data = D, lambda = 1)
}
