% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-base-model-function.R
\name{fit.strt_ev_tree}
\alias{fit.strt_ev_tree}
\title{Fit a stratified event tree}
\usage{
fit.strt_ev_tree(evt, data = NULL, lambda = 0)
}
\arguments{
\item{evt}{The stratified event tree object to be fitted}

\item{data}{the data.frame used to fit the event tree}

\item{lambda}{the laplace smoothing}
}
\value{
A stratified event tree object
}
\description{
Fit a stratified event tree
}
\details{
Distribute the counts along the event tree
}
