% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-model-selection.R
\name{full_indep}
\alias{full_indep}
\alias{full}
\alias{full.table}
\alias{full.data.frame}
\alias{indep}
\alias{indep.table}
\alias{indep.data.frame}
\title{Full and independent staged event tree}
\usage{
full(
  data,
  order = NULL,
  join_unobserved = FALSE,
  lambda = 0,
  name_unobserved = "UNOBSERVED"
)

\method{full}{table}(
  data,
  order = names(dimnames(data)),
  join_unobserved = FALSE,
  lambda = 0,
  name_unobserved = "UNOBSERVED"
)

\method{full}{data.frame}(
  data,
  order = colnames(data),
  join_unobserved = FALSE,
  lambda = 0,
  name_unobserved = "UNOBSERVED"
)

indep(
  data,
  order = NULL,
  join_unobserved = FALSE,
  lambda = 0,
  name_unobserved = "UNOBSERVED"
)

\method{indep}{table}(
  data,
  order = names(dimnames(data)),
  join_unobserved = FALSE,
  lambda = 0,
  name_unobserved = "UNOBSERVED"
)

\method{indep}{data.frame}(
  data,
  order = colnames(data),
  join_unobserved = FALSE,
  lambda = 0,
  name_unobserved = "UNOBSERVED"
)
}
\arguments{
\item{data}{data to create the model, data.frame or table.}

\item{order}{character vector, order of variables.}

\item{join_unobserved}{logical, if situations with zero observations should
be joined.}

\item{lambda}{smoothing coefficient.}

\item{name_unobserved}{name to pass to \code{\link{join_unobserved}}.}
}
\description{
Full and independent staged event tree
}
\details{
Functions to create full or independent staged tree models.
The full (or saturated) staged tree is the model where every
situation is in a different stage, and thus the model has the
maximum number of parameters.
Conversely, the independent staged tree (\code{indep}) assign
all the situation related to the same variable to the same
stage, thus it is equivalent to the independence factorization
\eqn{P(X1, X2, X3, ...) = P(X1)P(X2)P(X3)...} .
}
\examples{

######### full model

DD <- generate_xor_dataset(4, 100)
modfull <- full(DD, lambda = 1)

######### independence model (data.frame)
DD <- generate_xor_dataset(4, 100)
system.time(model <- indep(DD, lambda = 1))
model
}
