% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_toy_data.R
\name{generate_random_dataset}
\alias{generate_random_dataset}
\title{Generate a random binary dataset}
\usage{
generate_random_dataset(p, n)
}
\arguments{
\item{p}{number of variables.}

\item{n}{number of observations.}
}
\value{
A data.frame with \code{n} independent random variables.
}
\description{
Randomly generate a data.frame of independent binary variables.
}
\examples{
DD <- generate_random_dataset(p = 5, n = 1000)
}
