% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_toy_data.R
\name{generate_xor_dataset}
\alias{generate_xor_dataset}
\title{Generate a xor dataset}
\usage{
generate_xor_dataset(p, n, eps = 1.2)
}
\arguments{
\item{p}{number of variables.}

\item{n}{number of observations.}

\item{eps}{error.}
}
\value{
The xor dataset with \code{n} + 1 variables, where the first one is
the class variable \code{C} computed as a noisy xor.
}
\description{
Generate a xor dataset
}
\examples{
DD <- generate_xor_dataset(p = 5, n = 1000, eps = 1.2)
}
