% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-base-model-function.R
\name{has_ctables}
\alias{has_ctables}
\title{Check if the stages event tree has ctables field}
\usage{
has_ctables(object)
}
\arguments{
\item{object}{a staged event tree object}
}
\value{
logical
}
\description{
Check if the stages event tree has ctables field
}
\keyword{internal}
