% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages_em.R
\name{impute}
\alias{impute}
\title{impute data}
\usage{
impute(data, object, chain = FALSE)
}
\arguments{
\item{data}{a data.frame}

\item{object}{a sevt object}
}
\description{
Internal, No checks are performed
}
\keyword{internal}
