% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{inclusion.sevt}
\alias{inclusion.sevt}
\alias{inclusion.stages}
\title{Print inclusions of stages

Display the relationship between two staged tree models over the
same variables.}
\usage{
inclusion.sevt(object1, object2)

inclusion.stages(object1, object2)
}
\arguments{
\item{object1}{first staged event tree to compare}

\item{object2}{second staged event tree to compare}
}
\value{
a list with inclusion relations between stage
structures of each variable in the model
}
\description{
Print inclusions of stages

Display the relationship between two staged tree models over the
same variables.
}
\details{
\code{inclusion.sevt} computes the
inclusion/exclusion/equality/diversity between
the stages structures of the two models.

\code{inlcusion.stages} is the deprecated name of
\code{inclusion.sevt}, kept for back compatibility.
}
\examples{
mod1 <- bhc.sevt(full(PhDArticles[, 1:5], lambda = 1))
mod2 <- fbhc.sevt(full(PhDArticles[, 1:5], lambda = 1))
inclusion.sevt(mod1, mod2)
}
