% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{inclusion_stages}
\alias{inclusion_stages}
\title{Inclusions of stages

Display the relationship between two staged tree models over the
same variables.}
\usage{
inclusion_stages(object1, object2)
}
\arguments{
\item{object1}{first staged event tree to compare}

\item{object2}{second staged event tree to compare}
}
\value{
a list with inclusion relations between stage
structures of each variable in the model
}
\description{
Inclusions of stages

Display the relationship between two staged tree models over the
same variables.
}
\details{
Computes the
inclusion/exclusion/equality/diversity between
the stages structures of the two models.
}
\examples{
mod1 <- bhc(full(PhDArticles[, 1:5], lambda = 1))
mod2 <- fbhc(full(PhDArticles[, 1:5], lambda = 1))
inclusion_stages(mod1, mod2)
}
