% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{is_fitted.sevt}
\alias{is_fitted.sevt}
\title{Check if a staged event tree is fitted}
\usage{
is_fitted.sevt(x)
}
\arguments{
\item{x}{a staged event tree object}
}
\value{
logical
}
\description{
Check if a staged event tree is fitted
}
