% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{is_fitted_sevt}
\alias{is_fitted_sevt}
\title{Check if a staged event tree is fitted}
\usage{
is_fitted_sevt(object)
}
\arguments{
\item{object}{a staged event tree object.}
}
\value{
logical.
}
\description{
Check if a staged event tree is fitted
}
\keyword{internal}
