% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-model-selection.R
\name{join0stages}
\alias{join0stages}
\title{join all situations with 0 observations}
\usage{
join0stages(object)
}
\arguments{
\item{object}{a staged event tree object}
}
\value{
a staged event tree object where equal probabilities are in
the same stage
}
\description{
join all situations with 0 observations
}
