% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_stages.R
\name{join_ll_delta}
\alias{join_ll_delta}
\title{Log-likelihood change from joining two stages}
\usage{
join_ll_delta(p1, p2, lambda, k)
}
\arguments{
\item{p1}{probability vector of the first stage.}

\item{p2}{probability vector of the second stage.}

\item{lambda}{the smoothing parameter.}

\item{k}{the number of levels of the variable.}
}
\value{
a numeric scalar, the change in log-likelihood.
}
\description{
Compute the change in log-likelihood that \code{\link{join_stages_unsafe}}
would produce, without building the modified \code{sevt} object.
}
\details{
This must stay numerically identical to the log-likelihood update
performed in \code{\link{join_stages_unsafe}}; the two are exercised against
each other in the package tests. The change in degrees of freedom is not
returned because it is always \code{-(k - 1)}.
}
\keyword{internal}
