% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-custom-functions.R
\name{join_multiple_stages}
\alias{join_multiple_stages}
\title{Join multiple stages}
\usage{
join_multiple_stages(object, variable, join_stages)
}
\arguments{
\item{object}{an object of class \code{sevt}.}

\item{variable}{name of a variables that should be considered for the optimization}

\item{join_stages}{vector of stages to be merged.}
}
\value{
the staged event tree where all stages in \code{join_stages} are joined
}
\description{
Join multiple stages in a staged event tree object, updating
probabilities and log-likelihood accordingly.
}
\details{
This function joins all in one partition defined by the \code{part} vector,
updating probabilities and log-likelihood if
the object was fitted.
}
