% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_positions.R
\name{join_positions}
\alias{join_positions}
\title{join positions in a staged tree model}
\usage{
join_positions(model, var, s1, s2)
}
\arguments{
\item{model}{an object of class \code{sevt}.}

\item{var}{the name of a variable in the model.}

\item{s1}{stage to join}

\item{s2}{stage to join}
}
\description{
join positions in a staged tree model
}
\details{
this functions works similarly to the \code{join_stages}
function in the \code{stagedtrees} package, but it also joins
downstream stages to make nodes with stages \code{s1,s2} in the same
position. This function works properly only when downstream variables
from \code{var} have full stages vectors.
}
