% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-model-selection.R
\name{join_zero}
\alias{join_zero}
\title{Join situations with no observations}
\usage{
join_zero(object, fit = TRUE, trace = 0, name = NULL, lambda = object$lambda)
}
\arguments{
\item{object}{a fitted staged event tree}

\item{fit}{if the model's probability should be computed}

\item{trace}{if \code{> 0} print information to console}

\item{name}{string with a name for the new stage}

\item{lambda}{smoothing parameter for the fitting}
}
\value{
a staged event tree with situations with 0
observations merged in a single stage
}
\description{
Join situations with no observations
}
\details{
It takes as input a (fitted) staged event tree object
and looking at the \code{ctables} it joins, in the same stage, all the situations with zero
recorded observations. Since such joining does not change
the log-likelihood of the model, it is a useful (time-wise)
pre-processing before others model selection algorithms.
If \code{fit=TRUE} the model will be then re-fitted, if user sets
\code{fit=FALSE} the returned model will have no probabilities.
}
\examples{
DD <- generate_xor_dataset(n = 5, N = 10)
model_full <- full(DD, lambda = 1)
model <- join_zero(model_full)
logLik(model_full)
logLik(model)
BIC(model_full, model)
}
