% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-model-selection.R
\name{join_zero_counts}
\alias{join_zero_counts}
\title{Join situations with no observations}
\usage{
join_zero_counts(object, fit = TRUE)
}
\arguments{
\item{object}{a fitted staged event tree}

\item{fit}{if the probability should be re-computed}
}
\value{
a staged event tree with situations with 0 
observations merged
}
\description{
Join situations with no observations
}
