% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-util-functions.R
\name{KL_mat_prob}
\alias{KL_mat_prob}
\title{Compute the KL matrix}
\usage{
KL_mat_prob(x)
}
\arguments{
\item{x}{conditional probability table}
}
\value{
The matrix witht the symmetric KL (KL(i,j) + KL(j,i))
}
\description{
Compute the KL matrix
}
