% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-inference-functions.R
\name{logLik.staged_ev_tree}
\alias{logLik.staged_ev_tree}
\title{Compute log lik of a staged tree}
\usage{
\method{logLik}{staged_ev_tree}(object, data = NULL, ...)
}
\arguments{
\item{object}{the staged event tree object}

\item{data}{the dataset (default to `NULL``)}

\item{...}{additional parameters}
}
\description{
Compute log lik of a staged tree
}
