% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-inference-functions.R
\name{logLik.strt_ev_tree}
\alias{logLik.strt_ev_tree}
\title{Compute log lik of a stratified tree}
\usage{
\method{logLik}{strt_ev_tree}(object, ...)
}
\arguments{
\item{object}{the startified event tree object}

\item{...}{additional parameters}
}
\description{
Compute log lik of a stratified tree
}
\examples{

DD <- as.data.frame(sapply(1:5, function(i) {
                          return(as.factor(sample(c(1:3),
                          size=100, replace = TRUE)))
                     }))
sevt <- staged_ev_tree(DD, fit = TRUE)
evt <- strt_ev_tree(sevt)
logLik(evt)
}
