% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-inference-functions.R
\name{LR_test}
\alias{LR_test}
\title{Likelihood Ratio Test for staged trees models}
\usage{
LR_test(mod1, mod2)
}
\arguments{
\item{mod1}{simplest staged tree model.}

\item{mod2}{model that has to be nested in mod1. It must have a number of degrees of freedom
greater or equal to the one of \code{mod1}.}
}
\value{
It returns the corresponding p-value.
}
\description{
Likelihood Ratio Test for staged trees models
}
\details{
It computes the likelihood ratio test for two nested staged trees models.
The function checks automatically if the models are nested, returning an error if they are not.
}
\examples{
phd.mod1 <- stndnaming(stages_hc(indep(PhDArticles, order = order)))
phd.mod2 <- stndnaming(stages_hc(full(PhDArticles, order = order)))
LR_test(phd.mod1, phd.mod2)
}
