% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-base-model-function.R
\name{make_ctables}
\alias{make_ctables}
\title{Add data to a stratified event tree}
\usage{
make_ctables(object, data)
}
\arguments{
\item{object}{A stratified event tree, a list with a \code{tree} field}

\item{data}{table or data.frame containing observations
of the variable in \code{object}}
}
\value{
A list of \code{ftable}s.
}
\description{
Add the \code{ctables} field with data distributed along
the path of the tree.
}
\details{
Distribute the counts along the event tree.
This is an internal function, the user will
usually just directly fit the staged event tree
model using \code{sevt.fit}.
We refer here to stratified event tree, because actually
the stage information is never used and thus this function
will work for an object with only a \code{tree} field.
}
\keyword{internal}
