% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{new_label}
\alias{new_label}
\title{give a safe-to-add label that is not in \code{labels}
INTERNAL USE}
\usage{
new_label(labels)
}
\arguments{
\item{labels}{vector of labels (strings)}
}
\value{
a string label that is different from each \code{labels}
}
\description{
give a safe-to-add label that is not in \code{labels}
INTERNAL USE
}
