% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_toy_data.R
\name{noisy_xor}
\alias{noisy_xor}
\title{noisy xor function}
\usage{
noisy_xor(x, eps = 0)
}
\arguments{
\item{x}{a vector of +1 and -1.}

\item{eps}{the uniform noise amount.}
}
\value{
the computed noisy xor.
}
\description{
noisy xor function
}
\keyword{internal}
