% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{nvar.sevt}
\alias{nvar.sevt}
\title{Get the number of variables}
\usage{
nvar.sevt(x)
}
\arguments{
\item{x}{a staged event tree object}
}
\value{
integer, the number of variables
}
\description{
Get the number of variables
}
\examples{

mod <- indep(PhDArticles)
nvar.sevt(mod)
}
