% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{nvar_sevt}
\alias{nvar_sevt}
\title{Number of variables}
\usage{
nvar_sevt(object)
}
\arguments{
\item{object}{An object of class \code{sevt}.}
}
\value{
integer, the number of variables.
}
\description{
Utility returning the number of variables
in a staged event tree model.
}
\examples{

mod <- indep(PhDArticles)
nvar_sevt(mod)
}
