% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-custom-functions.R
\name{partition_stages}
\alias{partition_stages}
\title{Join stages of a staged event tree according to a partition}
\usage{
partition_stages(object, variable, stages, part)
}
\arguments{
\item{object}{an object of class \code{sevt}.}

\item{variable}{variable.}

\item{stages}{vector of stages considered for joining.}

\item{part}{vector of partitioning the stages into n bins.}
}
\value{
the staged event tree where the stages are joined according to
the partition vector
}
\description{
Join multiple stages in a staged event tree object, updating
probabilities and log-likelihood accordingly.
}
\details{
This function joins all in one partition defined by the \code{part} vector,
updating probabilities and log-likelihood if
the object was fitted.
}
