% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-inference-functions.R
\name{path_probability.sevt}
\alias{path_probability.sevt}
\title{Compute probability of a path from root}
\usage{
path_probability.sevt(object, x, log = FALSE)
}
\arguments{
\item{object}{a staged event tree object}

\item{x}{the path}

\item{log}{logical}
}
\value{
The probability of the given path or its logarithm if \code{log=TRUE}
}
\description{
Compute probability of a path from root
}
\examples{

#########
DD <- generate_random_dataset(5, 100)
model <- staged_ev_tree(DD, fit = TRUE, lambda = 1)
path_probability.sevt(model, c("1", "-1"))
path_probability.sevt(model, c("1", "-1", "1", "-1", "1"),
 log = TRUE)
}
