% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-inference-functions.R
\name{path_probability.staged_ev_tree}
\alias{path_probability.staged_ev_tree}
\title{Compute probability of a path}
\usage{
path_probability.staged_ev_tree(object, x, log = FALSE)
}
\arguments{
\item{object}{a staged event tree object}

\item{x}{the path}

\item{log}{logical}
}
\value{
The probability of the given path or its logarithm if \code{log=TRUE}
}
\description{
Compute probability of a path
}
