% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{PhDArticles}
\alias{PhDArticles}
\title{PhD Students Publications}
\format{
A data frame with 915 rows and 6 variables:
\describe{
\item{Articles}{Number of articles during the last 3 years of PhD: either
\code{0}, \code{1-2} or \code{>2}.}
\item{Gender}{\code{male} or \code{female}.}
\item{Kids}{\code{yes} if the student has at least one kid 5 or younger,
\code{no} otherwise.}
\item{Married}{\code{yes} or \code{no}.}
\item{Mentor}{Number of publications of the student's mentor:
\code{low} between 0 and 3, \code{medium} between 4 and 10,
\code{high} otherwise.}
\item{Prestige}{\code{low} if the student is at a low-prestige university,
\code{high} otherwise.}
}
}
\source{
The data has been modified from the \code{Rchoice} package.
}
\usage{
PhDArticles
}
\description{
Number of publications of 915 PhD biochemistry students
during the 1950’s and 1960’s.
}
\references{
Long, J. S. (1990). The origins of sex differences in science.
\emph{Social Forces}, 68(4), 1297-1316.
}
\keyword{datasets}
