% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_parentslist.R
\name{plot.parentslist}
\alias{plot.parentslist}
\title{igraph plotting of DAGs (parentslist)}
\usage{
\method{plot}{parentslist}(
  x,
  col = c(context = "red", `context-partial` = "orange", partial = "blue", local =
    "purple", total = "black"),
  ...
)
}
\arguments{
\item{x}{an R object of calss \code{parentslist}.}

\item{col}{specification for the colors of the edge type.}

\item{...}{additional parameters passed to \code{plot.igraph}.}
}
\description{
igraph plotting of DAGs (parentslist)
}
\details{
This function is a simple wrapper around
\pkg{igraph}'s \code{plot.igraph}.
The \code{parentslist} object is converted to an igraph object
with \code{\link{as_igraph}} and then passed to \code{plot.igraph}.
Additional arguments \code{...} are passed to the \pkg{igraph}
plotting function, allowing customization.
}
\examples{
 \dontrun{
 model <- stages_bhc(full(Titanic))
 pl <- as_parentslist(model, silent = TRUE)
 plot(model, edge.label = NA)
 }
}
