% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-plot-functions.R
\name{plot.strt_ev_tree}
\alias{plot.strt_ev_tree}
\title{plot a stratified event tree}
\usage{
\method{plot}{strt_ev_tree}(evt, rmax = 1, rmin = 0.1, step = 2, ...)
}
\arguments{
\item{x}{event tree object}
}
\description{
plot a stratified event tree
}
