% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{Pokemon}
\alias{Pokemon}
\title{Pokemon Go Users}
\format{
A data frame with 999 rows and 5 variables:
\describe{
\item{Use}{\code{Y} if the individual used the app, \code{N} otherwise}
\item{Age}{\code{>30} if the individual is older than 30,
\code{<=30} otherwise}
\item{Degree}{\code{Yes} if the individual completed a Higher
Education degree, \code{No} otherwise}
\item{Gender}{\code{Male} or \code{Female}}
\item{Activity}{\code{Yes} if the individual was physically active
(i.e. had a walk longer than 30 mins, went for a run or
had a bike ride to get some exercise) in the past week
before the experiment, \code{No} otherwise}
}
}
\source{
\url{https://osf.io/xy5g6/}
}
\usage{
Pokemon
}
\description{
Demographic information of a population of possible Pokemon Go users.
}
\references{
Gabbiadini, Alessandro, Christina Sagioglou, and Tobias Greitemeyer.
"Does Pokémon Go lead to a more physically active life style?."
\emph{Computers in Human Behavior} 84 (2018): 258-263.
}
\keyword{datasets}
