% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{as.character.parentslist}
\alias{as.character.parentslist}
\alias{print.parentslist}
\title{Print a parentslist object}
\usage{
\method{as.character}{parentslist}(x, ...)

\method{print}{parentslist}(x, ...)
}
\arguments{
\item{x}{an object of class \code{parentslist}.}

\item{...}{additional arguments for compatibility.}
}
\value{
\code{as.character.parentslist} returns a string
encoding the associated directed graph and eventually
the context specific and local-partial independence.
The encoding is similar to the one returned by
\code{\link{modelstring}} in package \pkg{bnlearn}
and package \pkg{deal}.
}
\description{
Nice print of a parentslist object
}
