% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{print.staged_ev_tree}
\alias{print.staged_ev_tree}
\title{Print a staged event tree}
\usage{
\method{print}{staged_ev_tree}(x, ...)
}
\arguments{
\item{x}{the staged event tree obejct}

\item{...}{additional parameters (compatibility)}
}
\value{
An invisible copy of \code{x}
}
\description{
Print a staged event tree
}
