% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-inference-functions.R
\name{prob.sevt}
\alias{prob.sevt}
\title{Compute probabilities for a staged event tree}
\usage{
prob.sevt(object, x, log = FALSE)
}
\arguments{
\item{object}{a (fitted) staged event tree object}

\item{x}{the vector or data.frame of observations}

\item{log}{logical, if \code{TRUE} log-probabilities are returned}
}
\value{
the probabilities for each observation
}
\description{
Compute probabilities for a staged event tree
}
