% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-util-functions.R
\name{probdist}
\alias{probdist}
\alias{lp}
\alias{kl}
\alias{tv}
\title{Distances between probabilities}
\usage{
lp(x, y, p = 2, ...)

kl(x, y, ...)

tv(x, y, ...)
}
\arguments{
\item{x}{vector of probabilities}

\item{y}{vector of probabilitites}

\item{p}{exponent in the \eqn{L^p} norm}

\item{...}{additional parameters for compatibility}
}
\value{
The distance between \code{p} and \code{q}
}
\description{
Distances between probabilities
}
\details{
Functions to compute distnces between probabilities:
\itemize{
\item \code{lp}: the \eqn{L^p} norm
\item \code{kl}: the symmetrized Kullback-Leibler divergence
\item \code{tv}: the total variation or \eqn{L^1} norm
}
}
