% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_tree.R
\name{ps_tree}
\alias{ps_tree}
\title{Ps-tree}
\usage{
ps_tree(object, outcome, treatment, ignore = object$name_unobserved)
}
\arguments{
\item{object}{a fitted object of class \code{sevt}.}

\item{outcome}{the outcome variable.}

\item{treatment}{the treatment variable.}

\item{ignore}{ignored stages.}
}
\description{
Build the ps-tree associated to a given staged event tree.
}
