% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_operations.R
\name{rename_stage}
\alias{rename_stage}
\title{Rename stage(s) in staged event tree}
\usage{
rename_stage(object, var, stage, new)
}
\arguments{
\item{object}{an object of class \code{sevt}.}

\item{var}{name of a variable in \code{object}.}

\item{stage}{name of the stage to be renamed.}

\item{new}{new name for the stage.}
}
\value{
a staged event tree object where stage \code{stage} has been
renamed to \code{new}.
}
\description{
Change the name of a stage in a staged event tree.
}
\details{
Renames a stage without merging. If \code{new} already exists
as a stage for variable \code{var}, an error is raised because renaming
onto an existing stage would silently overwrite its probability vector
with incorrect values. To merge two stages use \code{\link{join_stages}}.
}
