% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-sampling-functions.R
\name{sample.sevt}
\alias{sample.sevt}
\title{Sample from a staged event tree}
\usage{
sample.sevt(object, n = 1)
}
\arguments{
\item{object}{the staged event tree object}

\item{n}{number of observations to sample}
}
\value{
A data frame containing a sample of size \code{n}
}
\description{
Sample from a staged event tree
}
\details{
it samples \code{n} observations according to
the estimated transition probabilities (\code{object$prob}) from
the dataset on which the staged event tree \code{object} is fitted
}
\examples{
mod <- naive.sevt(full(PhDArticles, lambda = 1))
sample.sevt(mod, 10)
}
