% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-sampling-functions.R
\name{sample_sevt}
\alias{sample_sevt}
\title{Sample from a staged event tree}
\usage{
sample_sevt(object, nsim = 1, seed = NULL)
}
\arguments{
\item{object}{the staged event tree object}

\item{nsim}{number of observations to sample}

\item{seed}{an object specifying if and how the random number generator
should be initialized (‘seeded’). Either NULL or an integer
that will be used in a call to set.seed.}
}
\value{
A data frame containing a sample of size \code{n}
}
\description{
Generate a random sample from the distribution encoded
in a staged event tree object.
}
\details{
It samples \code{n} observations according to
the transition probabilities (\code{object$prob}) in the model.
}
\examples{
model <- naive_sevt(full(PhDArticles, lambda = 1))
sample_sevt(model, 10)
}
