% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{set_stage}
\alias{set_stage}
\title{Set stage to path}
\usage{
set_stage(object, path, stage)
}
\arguments{
\item{object}{an object of class \code{sevt}.}

\item{path}{Vector of the path.}

\item{stage}{stage to be assigned.}
}
\description{
Set stage to path
}
\keyword{internal}
