% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{sevt.fit}
\alias{sevt.fit}
\alias{fit.sevt}
\title{Fit a staged event tree}
\usage{
sevt.fit(sevt, data = NULL, lambda = 0, ...)

fit.sevt(sevt, data = NULL, lambda = 0, ...)
}
\arguments{
\item{sevt}{a staged event tree}

\item{data}{data.frame or stratified event tree}

\item{lambda}{smoothing parameter}

\item{...}{additional parameters}
}
\value{
a fitted staged event tree
}
\description{
Fit a staged event tree
}
\details{
\code{fit.sevt} is just a wrapper of \code{sevt.fit},
it is provided for compatibilities. (DEPRECATED)
}
\examples{

#########
model <- staged_ev_tree(list(X = c("good", "bad"),
                        Y = c("high", "low")))
D <- data.frame(X = c("good", "good", "bad"),
                Y = c("high", "low", "low"))
model.fit <- sevt.fit(model, data = D, lambda = 1)
}
