% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-ceg-functions.R
\name{sevt2ceg}
\alias{sevt2ceg}
\title{Creates the CEG starting from a staged tree}
\usage{
sevt2ceg(object)
}
\arguments{
\item{object}{a staged event tree}
}
\value{
a ceg object
}
\description{
Creates the CEG starting from a staged tree
}
\details{
A ceg object is a staged event tree object with
additional information on the positions.
}
\examples{
DD <- generate_xor_dataset(3, 100)
model <- bhc_sevt(full(DD))
ceg <- sevt2ceg(model)
ceg$positions
}
