% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-graph-conversion.R
\name{sevt2graph}
\alias{sevt2graph}
\alias{sevt2edges}
\alias{sevt2vert}
\alias{graph_from_sevt}
\title{sevt to graph structure}
\usage{
sevt2edges(x, ignore = x$name_unobserved)

sevt2vert(x, ignore = x$name_unobserved)

graph_from_sevt(x, ignore = x$name_unobserved)
}
\arguments{
\item{x}{an object of class \code{\link{sevt}}.}

\item{ignore}{vector of stages which will be ignored and excluded,
by default the name of the unobserved stages stored in
\code{x$name_unobserved}.}
}
\value{
for \code{sevt2edges}: the edges list corresponding
to the staged tree represented by \code{x}

for \code{graph_from_sevt}: a graph object from the
\pkg{igraph} package.
}
\description{
Obtain the graph representation of a staged tree, either as adjacency
matrix or as edges list or as \pkg{igraph} object.
}
\details{
This utility functions can be used to obtain the adjacency
matrix or the edge list structure
to build the staged tree using a graph package (e.g. \pkg{igraph}).
We also provide a function \code{graph_from_sevt} to build directly the
graph structure as in \pkg{igraph}.
This can be useful, for example to plot the staged tree (see the examples).
}
\examples{
\dontrun{
library(igraph)
library(ggplot2)
library(ggraph)
mod <- full(Titanic) |> stages_bhc()
g <- graph_from_sevt(mod)
ggraph(g, "sugiyama") + 
geom_edge_fan(aes( 
    label = label,
   label_pos = 0.5 + runif(length(label),-0.1,0.1)), 
   angle_calc = "along",show.legend = FALSE,check_overlap = FALSE,
   end_cap = circle(0.02, 'npc'),
   arrow = grid::arrow(angle = 25, 
                       length = unit(0.025, "npc"),
                       type = "closed")) +
 geom_node_point(aes(x = x, y = y, color = stage), size = 5, 
                 show.legend = FALSE) +
theme(aspect.ratio = 2) 
}
}
