% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sevt_fit_em.R
\name{sevt_fit_em}
\alias{sevt_fit_em}
\title{Fit a staged event tree with the hard Expectation-Maximization (EM) algorithm}
\usage{
sevt_fit_em(
  object,
  data = object$data_raw,
  lambda = NULL,
  scope = NULL,
  max_iter = 5,
  chain_impute = FALSE
)
}
\arguments{
\item{object}{an object of class \code{\link{sevt}}.}

\item{data}{data.frame with observations of
the variables in \code{object}.}

\item{lambda}{smoothing parameter. Default (NULL) to
lambda value stored in \code{object}. If no lambda value is
stored nor provided, 0 will be used with a warning.}

\item{scope}{which variable should be fitted. The value will be passed to
\code{\link{sevt_fit}}.}

\item{max_iter}{positive integer, the maximum number of iteration to be used
in the EM algorithm.}

\item{chain_impute}{logical. If \code{TRUE} chain predictions
(using the order of the variables in \code{object})
are used to impute
missing values. Otherwise, independent
predictions for each missing values are used.}
}
\value{
A fitted staged event tree,
See the return field of \code{\link{sevt_fit}}.
}
\description{
Estimate transition probabilities in a staged event tree from data containing
missing data using the hard (imputed) Expectation-Maximization algorithm.
}
