% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{sevt_nvar}
\alias{sevt_nvar}
\title{Number of variables}
\usage{
sevt_nvar(object)
}
\arguments{
\item{object}{An object of class \code{sevt}.}
}
\value{
integer, the number of variables.
}
\description{
Utility returning the number of variables
in a staged event tree model.
}
\keyword{internal}
