% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{sevt_varnames}
\alias{sevt_varnames}
\title{Variable names}
\usage{
sevt_varnames(object)
}
\arguments{
\item{object}{a staged event tree.}
}
\value{
A character vector.
}
\description{
Utility returning variable-names in a staged event tree
model.
}
\keyword{internal}
