% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-util-functions.R
\name{simple_clustering}
\alias{simple_clustering}
\title{perform a simple clustering in 2 groups}
\usage{
simple_clustering(M)
}
\arguments{
\item{M}{matrix of distance}
}
\value{
a list with component \code{I} and \code{J} (the two group of clusters)
}
\description{
clustering in 2 groups is performed starting
from a distance (positive, symmetric) matrix
}
