% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{split_stage_random}
\alias{split_stage_random}
\title{Split randomly a stage}
\usage{
split_stage_random(object, var, stage, p = 0.5)
}
\arguments{
\item{object}{a staged event tree object.}

\item{var}{the variable name.}

\item{stage}{the name of the stage.}

\item{p}{probability to move a situation from the
original stage into the new stage.}
}
\value{
a staged event tree object.
}
\description{
Randomly assign some of the paths to a new stage.
}
\details{
Splits randomly a given stage into two stages. More precisely,
it assigns each situation within the given stage into a new stage with
probability \code{p}.
}
\keyword{internal}
