% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{staged_ev_tree.bn.fit}
\alias{staged_ev_tree.bn.fit}
\title{Conversion from BN to staged event tree}
\usage{
\method{staged_ev_tree}{bn.fit}(x, ...)
}
\arguments{
\item{x}{\code{bn.fit} object}

\item{...}{additional parameters}
}
\description{
Conversion from BN to staged event tree
}
