% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{staged_ev_tree.data.frame}
\alias{staged_ev_tree.data.frame}
\title{Staged (stratified) event tree}
\usage{
\method{staged_ev_tree}{data.frame}(x, order = colnames(x), fit = FALSE,
  lambda = 0)
}
\arguments{
\item{x}{data.frame with the observation}

\item{order}{vector of order, default to the order in `x`}
}
\value{
the staged event tree object
}
\description{
Builds the staged event tree for a set of categorical variables,
the order can be specified.
}
