% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{staged_ev_tree.default}
\alias{staged_ev_tree.default}
\title{Staged (stratified) event tree}
\usage{
\method{staged_ev_tree}{default}(x, ...)
}
\arguments{
\item{x}{object that will be coerced to a data.frame}

\item{...}{additional parameters that will be passed
to the appropiate method, see \link{staged_ev_tree.data.frame}}
}
\value{
A staged event tree object
}
\description{
Builds the staged event tree for a set of categorical variables,
the order can be specified.
}
