% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{staged_ev_tree.fit}
\alias{staged_ev_tree.fit}
\title{Fit a stratified event tree}
\usage{
\method{staged_ev_tree}{fit}(sevt, data, lambda = 0, ...)
}
\arguments{
\item{sevt}{The staged event tree object to be fitted}

\item{data}{the data.frame used to fit the staged event tree}

\item{lambda}{the laplace smoothing}
}
\value{
A staged event tree object with the conditional probabilities fitted
}
\description{
Fit a stratified event tree
}
