% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{staged_ev_tree.list}
\alias{staged_ev_tree.list}
\title{Staged (stratified) event tree}
\usage{
\method{staged_ev_tree}{list}(x, ...)
}
\arguments{
\item{x}{list named as the variable and containing the vector of the levels
for each variable}

\item{...}{additional parameters}
}
\value{
The staged event tree object
}
\description{
Builds the staged event tree for a set of categorical variables,
the order can be specified.
}
\details{
The staged (stratified) event tree returned is the minimal one,
         that is the one with just one stage per variable (equivalent to a complete independent model).
}
