% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{staged_ev_tree}
\alias{staged_ev_tree}
\title{Staged (stratified) event tree}
\usage{
staged_ev_tree(x, ...)
}
\arguments{
\item{x}{data.frame, list or stratified event tree object}

\item{...}{additional parameters to be passed
to the appropiate method, see \link{staged_ev_tree.data.frame}}
}
\value{
A staged event tree object (see Details)
}
\description{
Builds the staged event tree for a set of categorical variables,
the order can be specified.
}
\details{
A staged event tree object consist
}
