% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{staged_ev_tree.strt_ev_tree}
\alias{staged_ev_tree.strt_ev_tree}
\title{Staged event tree}
\usage{
\method{staged_ev_tree}{strt_ev_tree}(x, ...)
}
\arguments{
\item{x}{A stratified event tree object}

\item{...}{additional parameters}
}
\value{
the equivalent staged event tree object
}
\description{
Staged event tree
}
\details{
The function expands the conditional probabilities into a full
          stratified event tree
}
