% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{staged_ev_tree.table}
\alias{staged_ev_tree.table}
\title{Staged (stratified) event tree}
\usage{
\method{staged_ev_tree}{table}(x, order = names(dimnames(x)),
  full = FALSE, fit = TRUE, ...)
}
\arguments{
\item{x}{table with the observation}

\item{order}{vector of order of the variables}

\item{full}{logical, if the full model should be built instead}

\item{fit}{logical}

\item{...}{additional parameters to be passed to other methods (mainly
\link{fit.staged_ev_tree})}
}
\value{
A staged event tree object
}
\description{
Builds the staged event tree from data provided as a \code{table}
}
\details{
The staged event trees
}
\examples{
data("Titanic")
model <- staged_ev_tree(Titanic, fit = TRUE, full = TRUE, lambda = 1)
plot(model)
}
