% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{stageinfo.sevt}
\alias{stageinfo.sevt}
\title{Print stage(s) info}
\usage{
stageinfo.sevt(object, var, stage = NULL)
}
\arguments{
\item{object}{a staged event tree object}

\item{var}{name of one variable}

\item{stage}{name(s) of stage for variable \code{var}}
}
\description{
Print stage(s) info
}
