% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{stages}
\alias{stages}
\title{Stages of a variable}
\usage{
stages(object, var = NULL)
}
\arguments{
\item{object}{a staged event tree.}

\item{var}{name of one variable in \code{object}.}
}
\value{
If \code{var} is specified, it returns a character vector with
stage names for the given variable
(that is \code{object$stages[[var]]}.
Otherwise, If \code{var} is not specified, \code{stages}
returns a list of character vectors containing the stages associated
to each variable in the model (that is \code{object$stages}).
}
\description{
Obtain the stages of a given variable in a staged event tree object.
}
