% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages_em.R
\name{stages_em}
\alias{stages_em}
\title{Structural EM for stages structure}
\usage{
stages_em(
  object,
  data = object$data_raw,
  alg_em = stages_bhc,
  max_iter_em = 5,
  incremental = FALSE,
  chain = FALSE,
  ...
)
}
\arguments{
\item{object}{an object of class sevt with fitted probabilities
and data, as returned by \code{\link{full}}
or \code{\link{sevt_fit}}.}

\item{data}{the data with missing values to be used in the EM algorithm.}

\item{alg_em}{a function that performs stages structure estimation. Similar to
\code{\link{stages_bhc}} or \code{\link{stages_hclust}}.
The function \code{alg} must accept the argument
\code{scope}.}

\item{max_iter_em}{the maximum number of iterations of the EM algorithm.}

\item{incremental}{logical. If \code{TRUE}, at each iteration of the EM algorithm
\code{alg_em} is initialized on the results of the
previous step. If \code{FALSE}, at each iteration
\code{alg_em} is initialized on the
initial argument \code{object}.}

\item{chain}{logical. If \code{TRUE} chain predictions are used to impute
missing values. Otherwise, independent
predictions for each missing values are used.}

\item{...}{additional arguments passed to \code{alg}.}
}
\value{
The final staged event tree obtained.
}
\description{
Structural EM for stages structure
}
