% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-model-selection.R
\name{stages_fbhc}
\alias{stages_fbhc}
\title{Fast backward hill-climbing}
\usage{
stages_fbhc(
  object = NULL,
  score = function(x) {     return(-BIC(x)) },
  max_iter = Inf,
  scope = NULL,
  ignore = object$name_unobserved,
  trace = 0
)
}
\arguments{
\item{object}{a staged event tree model}

\item{score}{the score function to be maximized}

\item{max_iter}{the maximum number of iteration}

\item{scope}{names of variables that should be considered for the optimization}

\item{ignore}{stage names that should be ignored}

\item{trace}{if >0 increasingly amount of info
is printed (via \code{message})}
}
\value{
The final staged event tree obtained.
}
\description{
Fast backward hill-climbing search of staged event trees with
iterative joining of stages.
}
\details{
For each variable the algorithm try to join stages
and move to the first model that increase the score. When no
increase is possible it moves to the next variable.
}
\examples{
DD <- generate_xor_dataset(n = 5, N = 100)
model <- stages_fbhc(full(DD), trace = 2)
summary(model)
}
