% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-custom-functions.R
\name{stages_full_ordered_search}
\alias{stages_full_ordered_search}
\title{TO DO: Add examples}
\usage{
stages_full_ordered_search(
  object,
  variable = NULL,
  n_init = NULL,
  n_bins = 3,
  score = function(x) {     return(-BIC(x)) },
  ignore = object$name_unobserved
)
}
\arguments{
\item{object}{an object of class \code{sevt} with fitted probabilities and
data, as returned by \code{full} or \code{sevt_fit}.}

\item{variable}{name of a variables that should be considered for the optimization.}

\item{n_init}{initial number of stages per subset considered for the optimization.}

\item{n_bins}{final number of stages.}

\item{score}{the score function to be maximized.}

\item{ignore}{vector of stages which will be ignored and left untouched,
by default the name of the unobserved stages stored in
\code{object$name_unobserved}.}
}
\value{
The final staged event tree obtained.
}
\description{
Full search of an ordered variable.
}
\details{
Full search on one level of a staged event tree with
iterative joining of "adjacent" stage for an ordered variable.

For the given variable the algorithm separates the stages into bins
and joins all stages in one bin. It searches all possible splittings of the
stages into the bins and returns the best scoring model.
}
