% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-custom-functions.R
\name{stages_ordered_bhc}
\alias{stages_ordered_bhc}
\title{TO DO: Add examples
Backward hill-climbing for ordered variables}
\usage{
stages_ordered_bhc(
  object,
  variable = NULL,
  n_init = NULL,
  score = function(x) {     return(-BIC(x)) },
  max_iter = Inf,
  ignore = object$name_unobserved,
  trace = 0
)
}
\arguments{
\item{object}{an object of class \code{sevt} with fitted probabilities and
data, as returned by \code{full} or \code{sevt_fit}.}

\item{variable}{name of a variables that should be considered for the optimization.}

\item{n_init}{initial number of stages per subset considered for the optimization.}

\item{score}{the score function to be maximized.}

\item{max_iter}{the maximum number of iterations per variable.}

\item{ignore}{vector of stages which will be ignored and left untouched,
by default the name of the unobserved stages stored in
\code{object$name_unobserved}.}

\item{trace}{if >0 increasingly amount of info
is printed (via \code{message}).}
}
\value{
The final staged event tree obtained.
}
\description{
Greedy search on one level of a staged event tree with
iterative joining of "adjacent" stages for an ordererd variable.
}
\details{
For the given variable the algorithm tries to join "adjacent" stages
and moves to the best model that increases the score. When no
increase is possible it terminates and returns the best scoring model.
}
