% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{stndnaming}
\alias{stndnaming}
\title{Standard renaming of stages}
\usage{
stndnaming(object, uniq = FALSE, prefix = FALSE, ignore = NULL)
}
\arguments{
\item{object}{a staged event tree.}

\item{uniq}{logical, if stage numbers should be unique over all tree.}

\item{prefix}{logical, if stage names should be prefixed with variable name.}

\item{ignore}{stages name which should be left untouched.}
}
\value{
a staged event tree object with stages named with
consecutive integers.
}
\description{
Rename all stages in a staged event tree.
}
\examples{
DD <- generate_xor_dataset(4, 100)
model <- fbhc(full(DD, join_zero = TRUE))
model$stages
model1 <- stndnaming(model)
model1$stages

### unique stage names in all tree
model2 <- stndnaming(model, uniq = TRUE)
model2$stages

### prefix stage names with variable name 
model3 <- stndnaming(model, prefix = TRUE)
model3$stages

### some stage names left untouched
model4 <- stndnaming(model, ignore = c("2", "4"), prefix = TRUE)
model4$stages
}
