% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-base-model-function.R
\name{strt_ev_tree.default}
\alias{strt_ev_tree.default}
\title{Stratified event tree}
\usage{
\method{strt_ev_tree}{default}(x, ...)
}
\arguments{
\item{x}{object to be coerced to data.frame}

\item{...}{additional parameters}
}
\description{
Stratified event tree
}
