% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-base-model-function.R
\name{strt_ev_tree.fit}
\alias{strt_ev_tree.fit}
\title{Fit a stratified event tree}
\usage{
\method{strt_ev_tree}{fit}(evt, data, lambda = 0)
}
\arguments{
\item{evt}{The stratified event tree object to be fitted}

\item{data}{the data.frame used to fit the event tree}

\item{lambda}{the laplace smoothing}
}
\description{
Fit a stratified event tree
}
