% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-base-model-function.R
\name{strt_ev_tree}
\alias{strt_ev_tree}
\title{Stratified event tree}
\usage{
strt_ev_tree(x, order = NULL, fit = FALSE, ...)
}
\arguments{
\item{x}{data.frame or list}

\item{order}{Vector of variables names, the order to build the event tree}

\item{fit}{If \code{TRUE} the contingency tables will be computed from \code{data}}

\item{...}{additional parameters}
}
\description{
Builds the complete stratified event tree for a set of categorical variables,
the order can be specified.
}
\examples{
a <- c(1,2)
}
