% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-base-model-function.R
\name{strt_ev_tree.table}
\alias{strt_ev_tree.table}
\title{Stratified event tree from a table}
\usage{
\method{strt_ev_tree}{table}(x, order = names(dimnames(x)), fit = FALSE, lambda = 0, ...)
}
\arguments{
\item{x}{table}

\item{order}{vector of order to build the tree}

\item{fit}{logical}

\item{lambda}{laplace smoothing parameter}

\item{...}{additional parameters}
}
\value{
A stratified event tree object
}
\description{
Stratified event tree from a table
}
