% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtrees.R
\name{subtree}
\alias{subtree}
\title{Extract subtree}
\usage{
subtree(object, path)
}
\arguments{
\item{object}{an object of class \code{sevt}.}

\item{path}{the path from root after which extract the subtree.}
}
\value{
A staged event tree object corresponding to the subtree.
}
\description{
Extract subtree
}
\details{
Returns the subtree of the staged event tree, starting from
\code{path}.
}
\examples{
DD <- generate_random_dataset(4, 100)
model <- sevt(DD, full = TRUE)
plot(model)
model1 <- subtree(model, path = c("-1", "1"))
plot(model1)
}
