% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{subtree.sevt}
\alias{subtree.sevt}
\title{Extract subtree}
\usage{
subtree.sevt(object, path)
}
\arguments{
\item{object}{a staged event tree object}

\item{path, }{the path after which extract the subtree}
}
\value{
the staged event tree object corresponding to the subtree
}
\description{
Extract subtree
}
