% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-plot-functions.R
\name{text.sevt}
\alias{text.sevt}
\title{Add text to a staged even tree plot}
\usage{
\method{text}{sevt}(x, y = ylim[1], limit = 10, xlim = c(0, 1),
  ylim = c(0, 1), ...)
}
\arguments{
\item{x}{staged event tree object}

\item{y}{the position of the labels}

\item{limit}{maximum number of variables plotted}

\item{xlim}{graphical parameter}

\item{ylim}{graphical parameter}

\item{...}{additional parameters}
}
\description{
Add text to a staged even tree plot
}
