% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-ceg-functions.R
\name{toCeg.sevt}
\alias{toCeg.sevt}
\title{Creates the CEG starting from a staged tree}
\usage{
toCeg.sevt(object)
}
\arguments{
\item{object}{a staged event tree}
}
\value{
a ceg object
}
\description{
Creates the CEG starting from a staged tree
}
