% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-util-functions.R
\name{uni_idx}
\alias{uni_idx}
\title{Unique id from named list}
\usage{
uni_idx(x)
}
\arguments{
\item{x}{a named list}
}
\value{
A named list with unique ids
}
\description{
Unique id from named list
}
\examples{
\dontrun{uni_idx(list(A = c(1,2,3), B = c(1,2,3))}
}
\keyword{internal}
