% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{variable.names.sevt}
\alias{variable.names.sevt}
\title{Variable Names of a Staged Event Tree}
\usage{
\method{variable.names}{sevt}(object, ...)
}
\arguments{
\item{object}{an object of class \code{sevt}.}

\item{...}{for compatibility.}
}
\description{
Simple utilities returning variable names.
}
