% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{varnames.sevt}
\alias{varnames.sevt}
\title{Get variable names}
\usage{
varnames.sevt(x)
}
\arguments{
\item{x}{a staged event tree object}
}
\value{
vector with variable names
}
\description{
Get variable names
}
\examples{

mod <-full(PhDArticles)
varnames.sevt(mod)
}
