% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{as_parentslist}
\alias{as_parentslist}
\title{Obtain the equivalent DAG list of parents}
\usage{
as_parentslist(x)
}
\arguments{
\item{x}{an object of class \code{sevt}}
}
\value{
An object of class \code{parentslist} for which a
print method exists.
Basically a list with
one entries for each variable with fields:
\itemize{
\item \code{values} the sample space for the variable.
\item \code{parents} The parents of the variable,
}
}
\description{
Compute, for each variable in the staged tree,
the parent set in the equivalent DAG.
}
\details{
The output of this function is an object of class
\code{parentslist} which is one of the possible encoding for
a directed graph.
If a context-specific or a local-partial independence is detected
a message is printed and the minimal super-model is returned.
}
\seealso{
\code{\link{print.parentslist}}.
}
