% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{as_sevt}
\alias{as_sevt}
\title{Convert to sevt class}
\usage{
as_sevt(x, order = NULL)
}
\arguments{
\item{x}{an R object to be coerced.}

\item{order}{order of the variables.}
}
\value{
the equivalent sevt object.
}
\description{
Convert to sevt class
}
\details{
Only a method for objects of class \code{bn.fit}
(\code{bnlearn} package) is available.
}
