% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-model-selection.R
\name{backward_hill_climb}
\alias{backward_hill_climb}
\title{backword hill-climbing}
\usage{
backward_hill_climb(object = NULL, score = function(x) return(-BIC(x)),
  max_iter = Inf, verbose = FALSE)
}
\arguments{
\item{object}{a staged event tree model}

\item{score}{the score function to be maximized}

\item{max_iter}{the maximum number of iteration}

\item{verbose}{If info should be printed (via \code{message})}
}
\description{
Each iter move to the best model, from more complicated to simpler model
}
