% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-model-selection.R
\name{backward_hill_climb}
\alias{backward_hill_climb}
\title{backword hill-climbing}
\usage{
backward_hill_climb(object = NULL, data = NULL, order = NULL,
  lambda = 1, score = function(x) return(-BIC(x)), eps = 1e-04,
  max_iter = 100, verbose = FALSE)
}
\arguments{
\item{object}{a staged event tree model (optional)}

\item{data}{the data (can be not specified if it is attached to \code{object})}

\item{order}{the order of the tree (optional)}

\item{lambda}{the laplace smoothing factor}

\item{score}{the score function to be maximized}

\item{eps}{the stopping criteria for the relative score increase}

\item{max_iter}{the maximum number of iteration}

\item{verbose}{If info should be printed}
}
\description{
Each iter move to the best model, from more complicated to simpler model
}
