% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-model-selection.R
\name{backward_joining_KL}
\alias{backward_joining_KL}
\title{Backword joining of stages}
\usage{
backward_joining_KL(object = NULL, thr = 0.01, trace = 0)
}
\arguments{
\item{object}{the staged event tree from where to start}

\item{thr}{the threshold for joining stages}

\item{trace}{if >0 increasingly amount of info 
is printed (via \code{message})}
}
\value{
the learned staged event tree
}
\description{
Join stages from more complex to simpler models
using entropy do decide if join or not a stage
}
\examples{
DD <- generate_xor_dataset(n = 5, N = 1000)
model_full <- staged_ev_tree(DD, fit = TRUE, full = TRUE, lambda = 1)
model <- backward_joining_KL(model_full, trace = 2)
BIC(model_full, model)
}
