% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-plot-functions.R
\name{barplot.sevt}
\alias{barplot.sevt}
\title{Barplot of stage probabilities}
\usage{
\method{barplot}{sevt}(
  height,
  var = variable.names(height)[1],
  ignore = height$name.unobserved,
  beside = TRUE,
  horiz = FALSE,
  legend.text = FALSE,
  col = NULL,
  xlab = ifelse(horiz, "probability", NA),
  ylab = ifelse(!horiz, "probability", NA),
  ...
)
}
\arguments{
\item{height}{An object of class \code{sevt}.}

\item{var}{name of a variable in \code{object}.}

\item{ignore}{array of stages name that should not be plotted.}

\item{beside}{a logical value. See \code{\link{barplot}}.}

\item{horiz}{a logical value. See \code{\link{barplot}}.}

\item{legend.text}{logical.}

\item{col}{color mapping for the stages, see \code{col}
argument in \code{\link{plot.sevt}}.}

\item{xlab}{a label for the x axis.}

\item{ylab}{a label for the y axis.}

\item{...}{additional arguments passed to \code{\link{barplot}}.}
}
\description{
Barplot of stage probabilities
}
\examples{
model <- stages_fbhc(full(PhDArticles, lambda = 1))
barplot(model, "Kids", beside = TRUE)
}
