% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-plot-functions.R
\name{barplot_stages}
\alias{barplot_stages}
\title{Barplot method for staged event trees}
\usage{
barplot_stages(object, var, legend.text = FALSE, col = NULL, ...)
}
\arguments{
\item{object}{a staged tree object}

\item{var}{name of varaibles in the model}

\item{legend.text}{logical}

\item{col}{color mapping for the stages, see \code{col}
argument in \code{\link{plot.sevt}}}

\item{...}{addittional arguments passed to \code{\link{barplot}}}
}
\description{
Barplot method for staged event trees
}
\examples{
model <- fbhc(full(PhDArticles, lambda = 1))
barplot_stages(model, "Kids", beside = TRUE)
}
