% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-ceg-functions.R
\name{ceg2adjmat}
\alias{ceg2adjmat}
\title{Ceg to adjmat of graph}
\usage{
ceg2adjmat(x)
}
\arguments{
\item{x}{the ceg object}
}
\value{
the adj matrix
}
\description{
Ceg to adjmat of graph
}
\details{
it is used to prepare the plot of the ceg.
It transforms the ceg object into an adjacency matrix.
}
\examples{
DD <- generate_xor_dataset(3, 100)
model <- bhc_sevt(full(DD))
ceg <- sevt2ceg(model)
ceg2adjmat(ceg)
}
