% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{has_ctables}
\alias{has_ctables}
\alias{has_prob}
\alias{is_fitted_sevt}
\alias{check_sevt}
\alias{check_tree}
\alias{check_stages}
\alias{check_sevt_prob}
\alias{check_sevt_ctables}
\alias{check_sevt_fit}
\alias{check_same_tree}
\alias{check_var_in}
\alias{check_scope}
\title{Check sevt objects}
\usage{
has_ctables(object)

has_prob(object)

is_fitted_sevt(object)

check_sevt(object, arg = rlang::caller_arg(object), call = rlang::caller_env())

check_tree(tree, arg, call)

check_stages(object, arg, call)

check_sevt_prob(
  object,
  arg = rlang::caller_arg(object),
  call = rlang::caller_env()
)

check_sevt_ctables(
  object,
  arg = rlang::caller_arg(object),
  call = rlang::caller_env()
)

check_sevt_fit(
  object,
  arg = rlang::caller_arg(object),
  call = rlang::caller_env()
)

check_same_tree(
  x,
  y,
  arg1 = rlang::caller_arg(x),
  arg2 = rlang::caller_arg(y),
  call = rlang::caller_env()
)

check_var_in(
  var,
  object,
  arg1 = rlang::caller_arg(var),
  arg2 = rlang::caller_arg(object),
  call = rlang::caller_env()
)

check_scope(
  scope,
  object,
  arg1 = rlang::caller_arg(scope),
  arg2 = rlang::caller_arg(object),
  call = rlang::caller_env()
)
}
\arguments{
\item{object}{an object of class sevt}

\item{arg}{passed arg name}

\item{call}{passed call}

\item{tree}{tree to be checked, a named list}

\item{x}{a staged event tree object.}

\item{y}{a staged event tree object.}

\item{arg1}{passed arg1 name}

\item{arg2}{passed arg2 name}

\item{var}{name of a variable to be checked.}

\item{scope}{the scope variables to be checked}
}
\value{
logical.

logical.

logical.
}
\description{
Check sevt objects
}
\keyword{internal}
