% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{compare.sevt}
\alias{compare.sevt}
\alias{hamming.sevt}
\alias{stagesdiff.sevt}
\title{Compare two staged event tree}
\usage{
compare.sevt(object1, object2, method = "naive", return.tree = FALSE,
  plot = FALSE, ...)

hamming.sevt(object1, object2, return.tree = FALSE)

stagesdiff.sevt(object1, object2)
}
\arguments{
\item{object1}{a staged event tree}

\item{object2}{a staged event tree}

\item{method}{method to compare staged event trees}

\item{return.tree}{logical, if \code{TRUE} the difference tree is returned}

\item{plot}{logical}

\item{...}{additional paraters to be passed to \code{\link{plot.sevt}}}
}
\value{
Logical or a difference tree (if \code{tree = TRUE})
}
\description{
Compare two staged event tree
}
\details{
\code{compare.sevt} test if the stage structure of two \code{sevt}
objects
is the same.
Three methods are available:
\itemize{
\item \code{naive} first applies \code{\link{stndnaming.sevt}} to both
objects and then simply compares the resulting \code{$stages} lists.
\item \code{hamming} uses the \code{hamming.sevt} function that try to map
stages in the different objects finding the few number of nodes that
must be changed to obtain the same structure.
\item \code{stages} uses the \code{stagesdiff.sevt} function that compare
stages to check if the same stage strucure is present in both models.
}

Setting \code{return.tree = TRUE} will return the stages
structure difference obtained with the selected method.

Functions \code{hamming.sevt} and \code{stagesdiff.sevt} can also be
used directly.
}
\examples{

#########
data("PhDArticles")
mod1 <- bhc.sevt(full(PhDArticles, lambda = 1))
mod2 <- fbhc.sevt(full(PhDArticles, lambda = 1))
compare.sevt(mod1, mod2)
compare.sevt(mod1, mod2, method = "hamming", plot = TRUE)

##########
m0 <- full(PhDArticles, fit = TRUE, lambda = 0)
m1 <- bhc.sevt(m0)
m1 <- stndnaming.sevt(m1) # it works also without the stndnaming of the stages.
m2 <- bj.sevt(m0, distance = tv, thr = 0.25)
m2 <- stndnaming.sevt(m2)
stagesdiff.sevt(m1, m2)
}
