% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{compare_stages}
\alias{compare_stages}
\alias{hamming_stages}
\alias{stagesdiff}
\title{Compare two staged event tree}
\usage{
compare_stages(
  object1,
  object2,
  method = "naive",
  return.tree = FALSE,
  plot = FALSE,
  ...
)

hamming_stages(object1, object2, return.tree = FALSE)

stagesdiff(object1, object2)
}
\arguments{
\item{object1}{a staged event tree.}

\item{object2}{a staged event tree.}

\item{method}{method to compare staged event trees. It can be: \code{"naive"},
\code{"hamming"} or \code{"stages"}.}

\item{return.tree}{logical, if \code{TRUE} the difference tree is returned.}

\item{plot}{logical.}

\item{...}{additional parameters to be passed to \code{\link{plot.sevt}}.}
}
\value{
\itemize{
\item \code{compare_stages}: if \code{return.tree = FALSE}, logical: \code{TRUE} if the two
models are exactly equal, otherwise \code{FALSE}.
Else If \code{return.tree = TRUE}, the differences between
the two trees, according to the selected \code{method}.
\item \code{hamming_stages}: if \code{return.tree = FALSE}, integer, the minimum
number of situations where the stage should be changed to obtain the same
models. If \code{return.tree = TRUE} a stages-like structure showing which
situations should be modified to obtain the same models.
\item \code{stagesdiff}: a stages-like structure marking the situations belonging
to stages which are not the same.
}
}
\description{
Compare two stages event tree, return the differences of the stages
structure and plot the difference tree. Three different methods to
compute the difference tree are available.
}
\details{
\code{compare_stages} tests if the stage structure of two \code{sevt}
objects
is the same.
Three methods are available:
\itemize{
\item \code{naive} first applies \code{\link{stndnaming}} to both
objects and then simply compares the resulting stages lists
(\code{stages(object1)} and \code{stages(object2)}).
\item \code{hamming} uses the \code{hamming_stages} function that try to map
stages in the different objects finding the few number of nodes that
must be changed to obtain the same structure.
\item \code{stages} uses the \code{stagesdiff} function that compare
stages to check if the same stage structure is present in both models.
}

Setting \code{return.tree = TRUE} will return the stages
structure difference obtained with the selected method.

With \code{plot = TRUE} the plot of the difference tree is obtained.

If \code{return.tree = FALSE} the logical output is the same for the
three methods and thus the \code{naive} method should be used
since it is computationally faster.

To use the \code{hamming} method, the package \code{clue}
must be installed.

Functions \code{\link{hamming_stages}} and \code{\link{stagesdiff}}
can also be used directly.
}
\examples{
data("PhDArticles")
mod1 <- bhc(full(PhDArticles[, 1:4], lambda = 1))
mod2 <- fbhc(full(PhDArticles[, 1:4], lambda = 1))
compare_stages(mod1, mod2)

##########
m0 <- full(PhDArticles[, 1:4], lambda = 0)
m1 <- bhc(m0)
m2 <- bj(m0, distance = "tv", thr = 0.25)
stagesdiff(m1, m2)
}
