% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-inference-functions.R
\name{confint.sevt}
\alias{confint.sevt}
\title{Confidence intervals for staged event tree parameters}
\usage{
\method{confint}{sevt}(
  object,
  parm,
  level = 0.95,
  sides = c("two.sided", "left", "right"),
  method = c("goodman", "wald", "waldcc", "wilson"),
  ignore = object$name_unobserved,
  ...
)
}
\arguments{
\item{object}{an object of class \code{sevt}.}

\item{parm}{a specification of which parameters are to be given
confidence intervals, either a vector of numbers
or a vector of names. If missing, all parameters are considered.}

\item{level}{the confidence level required.}

\item{sides}{a character string specifying the side of the confidence interval, must be one of "two.sided"
(default), "left" or "right".}

\item{method}{a character string specifing which method to use: "goodman", "wald", "waldcc" or "wilson".}

\item{ignore}{vector of stages which will be ignored,
by default the name of the unobserved stages stored in
\code{object$name_unobserved}.}

\item{...}{additional argument(s) for methods.}
}
\value{
An object of class \code{confint.sevt} containing the desired confidence intervals.
}
\description{
Confint method for class \code{sevt}.
}
\details{
Compute confidence intervals for staged event trees. By default,
it computes two-sided confidence intervals at 95\% level for the
whole tree given in \code{object} with the wald method.
}
\examples{
m1 <- stages_bj(full(PhDArticles), distance = "kullback", thr = 0.01)
confint(m1, "Prestige", level = 0.90, sides = "left")
confint(m1, "Married")
confint(m1, c("Married", "Kids"))
}
\references{
Goodman, L. A. (1965) On Simultaneous Confidence Intervals for Multinomial Proportions Technometrics, 7, 247-254.

Wald, A. Tests of statistical hypotheses concerning several parameters when the number of observations is large, Trans. Am. Math. Soc. 54 (1943) 426-482.

Wilson, E. B. Probable inference, the law of succession and statistical inference, J.Am. Stat. Assoc. 22 (1927) 209-212.
}
\author{
The function is partially inspired by code in the
\code{MultinomCI} function from the \pkg{DescTools} package,
implemented by Andri Signorelli and Pablo J. Villacorta Iglesias.
}
