% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{df.sevt}
\alias{df.sevt}
\title{Get number of parameters of the model}
\usage{
df.sevt(x)
}
\arguments{
\item{x}{a staged event tree object}
}
\value{
integer, degrees of freedom of the staged event tree
}
\description{
Get number of parameters of the model
}
\examples{

#########
mod_f <- full(PhDArticles)
df.sevt(mod_f)

mod_i <- indep(PhDArticles)
df.sevt(mod_i)
}
