% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-util-functions.R
\name{distance_mat_stages}
\alias{distance_mat_stages}
\title{Compute the distance matrix}
\usage{
distance_mat_stages(x, distance = kl, ...)
}
\arguments{
\item{x}{list of conditional probabilities for each stage}

\item{distance}{the distance function e.g. \code{\link{kl}}}

\item{...}{additional parameters to be passed to the distance function}
}
\value{
The matrix witht the distances between stages
}
\description{
Compute the distance matrix
}
