% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{effindep.staged_ev_tree}
\alias{effindep.staged_ev_tree}
\title{Built (efficently) an independent event tree}
\usage{
effindep.staged_ev_tree(x, lambda, ctables = TRUE)
}
\arguments{
\item{x}{data}

\item{lambda}{the smoothing parameter}
}
\description{
Built (efficently) an independent event tree
}
