% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-util-functions.R
\name{find_stage}
\alias{find_stage}
\title{find the stage in the path}
\usage{
find_stage(object, path)
}
\arguments{
\item{object}{a staged event tree object}

\item{path}{vector of the path}
}
\value{
the stage name corresponding of the path
}
\description{
no checking is done
}
\examples{
\dontrun{
data("PhDArticles")
mod <- full(PhDArticles)
find_stage(mod, c("0", "male", "yes"))
find_stage(mod, c("0", "male", "no"))}
}
\keyword{internal}
