% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_indexing.R
\name{find_stage}
\alias{find_stage}
\title{Find the stage of the path}
\usage{
find_stage(object, path)
}
\arguments{
\item{object}{a staged event tree object.}

\item{path}{vector of the path.}
}
\value{
the stage name corresponding of the path.
}
\description{
no checking is done.
}
\keyword{internal}
