% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{fit.sevt}
\alias{fit.sevt}
\title{#' conversion
#' @param x Adj matrix
#' @param tree variables tree
#' @param ...
staged_ev_tree.matrix <- function(x, tree, ...){
  order <- gRbase::topoSort(adj_mat)
  object <- staged_ev_tree(tree[order])
  
}
Fit a staged event tree}
\usage{
fit.sevt(sevt, data = NULL, lambda = 0, ...)
}
\arguments{
\item{sevt}{The staged event tree object to be fitted}

\item{data}{the data.frame used to fit the staged event tree}

\item{lambda}{the laplace smoothing}

\item{...}{additional parameters}
}
\value{
A staged event tree object with the conditional probabilities fitted
}
\description{
#' conversion
#' @param x Adj matrix
#' @param tree variables tree
#' @param ...
staged_ev_tree.matrix <- function(x, tree, ...){
  order <- gRbase::topoSort(adj_mat)
  object <- staged_ev_tree(tree[order])
  
}
Fit a staged event tree
}
\examples{
model <- staged_ev_tree(list(X = c("good", "bad"), Y = c("high", "low")))
D <- data.frame(X = c("good", "good", "bad"),
                Y = c("high", "low", "low"))
model.fit <- fit.sevt(model, data = D, lambda = 1)
}
