% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-model-selection.R
\name{full_indep}
\alias{full_indep}
\alias{full}
\alias{full.default}
\alias{indep}
\alias{indep.default}
\alias{indep.data.frame}
\title{Full and independent staged event tree}
\usage{
full(data, join_zero = FALSE, name.join = "NA", lambda = 0)

\method{full}{default}(data, join_zero = FALSE, name.join = "NA", lambda = 0)

indep(data, join_zero = FALSE, name.join = "NA", lambda = 0)

\method{indep}{default}(data, join_zero = FALSE, name.join = "NA", lambda = 0)

\method{indep}{data.frame}(data, join_zero = FALSE, name.join = "NA", lambda = 0)
}
\arguments{
\item{data}{data to create the model, data.frame or table}

\item{join_zero}{logical, if situations with zero observations should
be joined}

\item{name.join}{name to pass to \code{\link{join_zero}}}

\item{lambda}{smoothing coefficient}
}
\description{
Full and independent staged event tree
}
\details{
full and indep models
}
\examples{

######### full model

DD <- generate_xor_dataset(4, 100)
modfull <- full(DD, lambda = 1)

######### independence model (data.frame)
DD <- generate_xor_dataset(4, 100)
system.time(model <- indep(DD, lambda = 1))
model
}
