% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-util-functions.R
\name{generate_random_dataset}
\alias{generate_random_dataset}
\title{generate a random binary dataset}
\usage{
generate_random_dataset(n = 2, N = 10000)
}
\arguments{
\item{n}{number of variables}

\item{N}{number of observations}
}
\value{
A data.frame with \code{n} independent random variables and
one class variable \code{C} computed as
\code{sign(sum(x * alpha) + runif(1, -eps, eps) + gamma)}
}
\description{
generate a random binary dataset
}
\examples{
DD <- generate_random_dataset(n = 5, 1000)
}
