% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-util-functions.R
\name{generate_xor_dataset}
\alias{generate_xor_dataset}
\title{generate a xor dataset}
\usage{
generate_xor_dataset(n = 2, N = 100, eps = 1.2)
}
\arguments{
\item{n}{number of variables}

\item{N}{number of observations}

\item{eps}{error}
}
\value{
The xor dataset
}
\description{
generate a xor dataset
}
\examples{
DD <- generate_xor_dataset(n = 5, N = 1000, eps = 1.2)
}
