% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-base-model-function.R
\name{has_prob}
\alias{has_prob}
\title{Check if the stages event tree has probabilitites}
\usage{
has_prob(object)
}
\arguments{
\item{object}{a staged event tree object.}
}
\value{
logical.
}
\description{
Check if the stages event tree has probabilitites
}
\keyword{internal}
