% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-model-selection.R
\name{hc.sevt}
\alias{hc.sevt}
\title{Hill-Climb Score optimization}
\usage{
hc.sevt(
  object,
  score = function(x) {     return(-BIC(x)) },
  max_iter = Inf,
  ignore = NULL,
  trace = 0
)
}
\arguments{
\item{object}{a staged event tree object}

\item{score}{a function that score staged event tree objects}

\item{max_iter}{the maximum number of iterations per variable}

\item{ignore}{a vector of stage names which will not be changed}

\item{trace}{integer, if positive information on the progress is
printed to console}
}
\value{
The final staged event tree obtained.
}
\description{
Hill-climbing search of staged event trees with
iterative moving of nodes between stages.
}
\details{
For each variable node-moves that best increase the
score are performed until no increase is possible.
A node-move is either changing the stage
associate to a node or move the node to a new stage.

The \code{ignore} argument can be used to specify stages that should not
be affected during the search, that is left untouched.
This is usefult for preserving structural zeroes and to speed-up
computations.
}
\examples{
model <- hc.sevt(full(PhDArticles[, 1:3], lambda = 1))
summary(model)

## preserve zero stages
start <- join_zero(indep(PhDArticles[,1:5]), name = "NA")
model <- hc.sevt(start, ignore = "NA")
}
