% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-model-selection.R
\name{hc.sevt}
\alias{hc.sevt}
\title{Hill-Climb Score optimization}
\usage{
hc.sevt(object, score = function(x) return(-BIC(x)), max_iter = Inf,
  trace = 0)
}
\arguments{
\item{object}{a staged event tree object}

\item{score}{a function that score staged event tree objects}

\item{max_iter}{the maximum number of iterations per variable}

\item{trace}{integer, if positive information on the progress is
printed to console}
}
\value{
A staged event tree object, the output of the optimization
}
\description{
Hill-Climb Score optimization
}
\details{
TO WRITE
}
