% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{inclusion.stages}
\alias{inclusion.stages}
\title{Inclusion relations between stage structures of two models estimated on the same dataset}
\usage{
inclusion.stages(object1, object2)
}
\arguments{
\item{object1}{first staged event tree to compare}

\item{object2}{second staged event tree to compare}
}
\value{
list with inclusion relations between stage structures of each variable of the dataset
}
\description{
Inclusion relations between stage structures of two models estimated on the same dataset
}
\examples{
mod1 <- stndnaming.sevt(bhc.sevt(full(PhDArticles, lambda = 1)))
mod2 <- stndnaming.sevt(fbhc.sevt(full(PhDArticles, lambda = 1)))
comparison <- inclusion.stages(mod1, mod2)
print(comparison)
}
