% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{join_stages}
\alias{join_stages}
\title{Join stages}
\usage{
join_stages(sevt, v, s1, s2)
}
\arguments{
\item{sevt}{staged event tree}

\item{v}{variable}

\item{s1}{first stage}

\item{s2}{second stage}
}
\value{
the staged event tree where \code{s1} and \code{s2} are joined
}
\description{
Join two stages in a staged event tree object, updating
probabilities and log-likelihood accordingly.
}
