% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-inference-functions.R
\name{logLik.sevt}
\alias{logLik.sevt}
\title{Compute log lik of a staged tree}
\usage{
\method{logLik}{sevt}(object, ...)
}
\arguments{
\item{object}{the staged event tree object}

\item{...}{additional parameters}
}
\description{
Compute log lik of a staged tree
}
\examples{

#########
data("Trump")
mod <- staged_ev_tree(Trump)
logLik(mod)
}
