% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-inference-functions.R
\name{logLik.sevt}
\alias{logLik.sevt}
\title{Log-Likelihood of a staged event tree}
\usage{
\method{logLik}{sevt}(object, ...)
}
\arguments{
\item{object}{the staged event tree object}

\item{...}{additional parameters}
}
\value{
An object of class \code{\link{logLik}}.
}
\description{
Compute, or extract the log-likelihood of a staged event tree.
}
\examples{
data("PhDArticles")
mod <- indep(PhDArticles)
logLik(mod)
}
