% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-inference-functions.R
\name{logLik.sevt}
\alias{logLik.sevt}
\title{Log-Likelihood of a staged event tree}
\usage{
\method{logLik}{sevt}(object, ...)
}
\arguments{
\item{object}{an object of class \code{sevt}
with fitted probabilities and associated data, as returned by \code{\link{sevt_fit}},
\code{\link{full}}, \code{\link{indep}} or one of the model selection algorithms.}

\item{...}{additional parameters (compatibility).}
}
\value{
An object of class \code{\link{logLik}}.
}
\description{
Compute, or extract the log-likelihood of a staged event tree.
}
\examples{
data("PhDArticles")
mod <- indep(PhDArticles)
logLik(mod)
}
