% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-inference-functions.R
\name{logLik.strt_ev_tree}
\alias{logLik.strt_ev_tree}
\title{Compute log lik of a stratified tree}
\usage{
\method{logLik}{strt_ev_tree}(object, data, ...)
}
\arguments{
\item{object}{the startified event tree object}

\item{data}{the dataset (default to `NULL``)}
}
\description{
Compute log lik of a stratified tree
}
