% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-model-selection.R
\name{naive.sevt}
\alias{naive.sevt}
\title{Naive staged event tree}
\usage{
naive.sevt(
  object,
  distance = kl,
  k = length(object$tree[[1]]),
  method = "complete",
  ignore = NULL,
  limit = length(object$tree),
  ...
)
}
\arguments{
\item{object}{a full staged event tree}

\item{distance}{a distance between probabilities}

\item{k}{number of clusters, that is stages per variable}

\item{method}{the agglomeration method to be used in \code{\link{hclust}}}

\item{ignore}{vector of stages which will be left untouched}

\item{limit}{the maximum number of variable to consider}

\item{...}{additional arguments to be passeed to \code{distance}}
}
\value{
A staged event tree with \code{k} stages per variable
}
\description{
Build a stage event tree with \code{k} stages for each variable
}
\details{
Old implementation which allows arbitrary distance
function, see \code{\link{hclust.sevt}} for the
most recent implementation.
}
\examples{
DD <- generate_xor_dataset(n = 4, N = 1000)[, 5:1]
naive_model <- naive.sevt(full(DD, lambda = 1))
pr <- predict(naive_model, newdata = DD[501:1000, ])
table(pr, DD$C[501:1000])
}
