% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-util-functions.R
\name{new_label}
\alias{new_label}
\title{New label}
\usage{
new_label(labels)
}
\arguments{
\item{labels}{vector of labels}
}
\value{
a string label that is different from each \code{labels}
}
\description{
give a safe-to-add label that is not in \code{labels}
}
\keyword{internal}
