% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-plot-functions.R
\name{node}
\alias{node}
\title{Plot a node}
\usage{
node(x, label = "", col = "black", cex_label = 1, cex_node = 1, ...)
}
\arguments{
\item{x}{the center}

\item{label}{the label}

\item{col}{color}

\item{cex_label}{cex parameter to be passed to text}

\item{cex_node}{cex parameter for nodes}

\item{...}{additional parameters passed to \code{par()}}
}
\description{
Plot a node
}
\keyword{internal}
