% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-plot-functions.R
\name{node}
\alias{node}
\title{Plot a node}
\usage{
node(x, r, label = "", col = "black", cex.label = 2 * sqrt(r), ...)
}
\arguments{
\item{x}{the center}

\item{label}{the label}

\item{col}{color}

\item{cex.label}{cex parameter to be passed to text}

\item{...}{additional parameters passed to \code{par()}}

\item{the}{radius}
}
\description{
Plot a node
}
