% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-plot-functions.R
\name{plot.sevt}
\alias{plot.sevt}
\title{plot a staged event tree}
\usage{
\method{plot}{sevt}(x, limit = 10, radius = 0.05, xlim = c(0, 1),
  ylim = c(0, 1), asp = 1, cex.label = 1, ...)
}
\arguments{
\item{x}{staged event tree object}

\item{limit}{maximum number of variables plotted}

\item{radius}{graphical parameter}

\item{xlim}{graphical parameter}

\item{ylim}{graphical parameter}

\item{asp}{graphical parameter}

\item{cex.label}{graphical parameter}

\item{...}{additional graphical parameters}
}
\description{
plot a staged event tree
}
