% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-plot-functions.R
\name{plot.sevt}
\alias{plot.sevt}
\title{plot a staged event tree}
\usage{
\method{plot}{sevt}(x, limit = 10, xlim = c(0, 1), ylim = c(0, 1),
  asp = 1, cex.label.nodes = 1, cex.label.edges = 1, cex.nodes = 2,
  col = NULL, ...)
}
\arguments{
\item{x}{staged event tree object}

\item{limit}{maximum number of variables plotted}

\item{xlim}{graphical parameter}

\item{ylim}{graphical parameter}

\item{asp}{graphical parameter}

\item{cex.label.nodes}{graphical parameter}

\item{cex.label.edges}{graphical parameter}

\item{cex.nodes}{graphical parameter}

\item{col}{color mapping for the stages, a named list with
names equal to the variables names in the model and
vectors named with stages names as components; otherwise
if \code{col == "stages"} the stage names will be used as
colors; otherwise if \code{col} is a function it will take
as input a vector of stages and output the corresponding colors.}

\item{...}{additional graphical parameters to be passed to
\code{points}, \code{lines}, \code{title},
\code{text} and \code{plot.window}.}
}
\description{
plot a staged event tree
}
