% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-plot-functions.R
\name{plot.sevt}
\alias{plot.sevt}
\title{Plot method for staged event trees}
\usage{
\method{plot}{sevt}(
  x,
  limit = 10,
  xlim = c(0, 1),
  ylim = c(0, 1),
  main = NULL,
  sub = NULL,
  asp = 1,
  cex.label.nodes = 1,
  cex.label.edges = 1,
  cex.nodes = 2,
  col = NULL,
  col.edges = "black",
  ignore = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class \code{sevt}.}

\item{limit}{maximum number of variables plotted.}

\item{xlim}{the x limits (x1, x2) of the plot.}

\item{ylim}{the y limits of the plot.}

\item{main}{an overall title for the plot.}

\item{sub}{a sub title for the plot.}

\item{asp}{the y/x aspect ratio.}

\item{cex.label.nodes}{amount of scaling for the node labels.
If set to \code{0} node labels are not showed.}

\item{cex.label.edges}{amount of scaling for the edge labels.
If set to \code{0} edge labels are not showed.}

\item{cex.nodes}{amount of scaling for the nodes of the tree.}

\item{col}{color mapping for the stages, a named list with
names equal to the variables names in the model and
vectors named with stages names as components; otherwise
if \code{col == "stages"} the stage names will be used as
colors; otherwise if \code{col} is a function it will take
as input a vector of stages and output the corresponding colors.}

\item{col.edges}{color for the edges.}

\item{ignore}{array of stages name that should not be plotted.}

\item{...}{additional graphical parameters to be passed to
\code{points}, \code{lines}, \code{title},
\code{text} and \code{plot.window}.}
}
\description{
\code{plot.sevt} is the plot method for staged event tree
objects. It allows easy plotting of staged event tree with some
options, mainly different ways to specify colors for the stages (see
examples).
}
\examples{

data("PhDArticles")
mod <- bj(full(PhDArticles, join_zero = TRUE))

### simple plotting
plot(mod)

### removing lables from nodes and edges and fill nodes
plot(mod, cex.label.nodes = 0, cex.label.edges = 0, pch = 16)

### reduce nodes size
plot(mod, cex.nodes = 1)

### change line width and nodes style
plot(mod, lwd = 3, pch = 5)

### changing palette
plot(mod, col = function(s) heat.colors(length(s)))

### or changing global palette
palette(hcl.colors(10, "Harmonic"))
plot(mod)

### avoid plotting NA stages
plot(mod, ignore = "NA")

### manually give stages colors
simple <- stages_hclust(full(PhDArticles, lambda = 1), k = 2)
#### simple has 2 stages per variable "1" and "2"
col <- lapply(simple$stages, function(s) {
  c("1" = "purple", "2" = "cyan")
})
plot(simple, col = col)
}
