% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-plot-functions.R
\name{plot.staged_ev_tree}
\alias{plot.staged_ev_tree}
\title{plot a staged event tree}
\usage{
\method{plot}{staged_ev_tree}(x, rmax = 1, rmin = 0.1, step = 2,
  limit = 10, ...)
}
\arguments{
\item{x}{event tree object}

\item{rmax}{graphical parameter}

\item{rmin}{graphical parameter}

\item{step}{graphical parameter}

\item{limit}{maximum number of levels plotted}

\item{...}{additional graphical parameters}
}
\description{
plot a staged event tree
}
