% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential_outcomes.R
\name{potential_outcomes}
\alias{potential_outcomes}
\title{Potential Outcomes}
\usage{
potential_outcomes(object, outcome, treatment)
}
\arguments{
\item{object}{a fitted object of class \code{sevt}.}

\item{outcome}{the outcome variable.}

\item{treatment}{the treatment variable.}
}
\value{
a matrix with potential outcomes.
}
\description{
Compute potential outcomes by randomizing
the treatment variable.
}
\examples{
model <- full(Titanic) |> stages_bhc()
potential_outcomes(model, "Survived", "Class", "pstree")
}
