% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential_outcomes.R
\name{potential_outcomes}
\alias{potential_outcomes}
\title{Potential Outcomes}
\usage{
potential_outcomes(
  object,
  outcome,
  treatment,
  method = c("randomization", "pstree")
)
}
\arguments{
\item{object}{a fitted object of class \code{sevt}.}

\item{outcome}{the outcome variable.}

\item{treatment}{the treatment variable.}

\item{method}{string, either \code{"randomization"} or \code{"pstree"}.}
}
\value{
a matrix with potential outcomes.
}
\description{
Compute potential outcomes by either randomizing the treatment variable
or via propensity score stratification.
}
\examples{
model <- full(Titanic) |> stages_bhc()
potential_outcomes(model, "Survived", "Class", "pstree")
}
