% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-classifier-model-functions.R
\name{predict.staged_ev_tree}
\alias{predict.staged_ev_tree}
\title{Predcict method for staged event tree}
\usage{
\method{predict}{staged_ev_tree}(object, newdata = NULL, class = NULL,
  prob = FALSE, ...)
}
\arguments{
\item{object}{A staged event tree}

\item{newdata}{The newdata to perform predictions}

\item{...}{additional parameters, see details}
}
\value{
A vector of predicitons
}
\description{
Predcict method for staged event tree
}
