% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-classifier-model-functions.R
\name{predict.staged_ev_tree}
\alias{predict.staged_ev_tree}
\title{Predcict method for staged event tree}
\usage{
\method{predict}{staged_ev_tree}(object, newdata = NULL, class = NULL,
  prob = FALSE, ...)
}
\arguments{
\item{object}{A staged event tree}

\item{newdata}{The newdata to perform predictions}

\item{class}{A string indicating the name of the variable to use as
the class variable}

\item{prob}{logical, if \code{TRUE} the probabilities of class are returned}

\item{...}{additional parameters, see details}
}
\value{
A vector of predictions or the corresponding probabilities

A vector of predicitons
}
\description{
Predcict method for staged event tree
}
\examples{
DD <- generate_xor_dataset(n = 5, 1000)
train <- DD[1:500,]
test <- DD[501:1000,]
order <- c("C", "X1", "X2", "X3", "X4", "X5")
model <- staged_ev_tree(train, order = order, full = TRUE, 
fit = TRUE, lambda = 1)
model <- backward_hill_climb(model)
pr <- predict(model, class = "C", newdata = test)
table(pr, test$C)
predict(model, class = "C", newdata = test, prob = TRUE)
}
