% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-random-sevt-functions.R
\name{random_parentslist}
\alias{random_parentslist}
\title{Generate a random \code{parentslist} object (DAG)}
\usage{
random_parentslist(n, k = 2, maxp = n)
}
\arguments{
\item{n}{number of variables.}

\item{k}{maximum number of levels for each variable.}

\item{maxp}{maximum cardinalty of parents sets.}
}
\value{
a \code{\link[=as_parentslist]{parentslist}} object.
}
\description{
generate a random DAG coded as
\code{\link[=as_parentslist]{parentslist}} object.
}
