% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{rename_stage}
\alias{rename_stage}
\title{Rename stage(s) in staged event tree}
\usage{
rename_stage(object, var, stage, new)
}
\arguments{
\item{object}{an object of class \code{sevt}.}

\item{var}{name of a variable in \code{object}.}

\item{stage}{name of the stage to be renamed.}

\item{new}{new name for the stage.}
}
\value{
a staged event tree object where stages \code{stage}
have been renamed to \code{new}.
}
\description{
Change the name of a stage in a staged event tree.
}
\details{
No internal checks are performed and as side effect
stages can be joined, if e.g. \code{new} is equal to the name
of a stage for variable \code{var}.
}
