% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{rename_stage}
\alias{rename_stage}
\title{rename stage(s) in staged even tree}
\usage{
rename_stage(object, var, stage, new.label)
}
\arguments{
\item{object}{staged event tree}

\item{var}{string, one of the variable in
the staged tree}

\item{stage}{string or vector, the name
of the stage(s) to be renamed}

\item{new.label}{new name for the stage(s)}
}
\value{
a staged event tree object where stages \code{stage}
have been renamed to \code{new.label}
}
\description{
rename stage(s) in staged even tree
}
