% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-sampling-functions.R
\name{sample.sevt}
\alias{sample.sevt}
\title{Sample from a staged event tree}
\usage{
sample.sevt(object, n = 1)
}
\arguments{
\item{object}{the staged event tree object}

\item{n}{number of observations}
}
\value{
A data frame containing a sample of size \code{n}
}
\description{
Sample from a staged event tree
}
\details{
TO WRITE
}
\examples{

#########
data("Titanic")
mod <- naive.sevt(full(Titanic, lambda = 1))
sample.sevt(mod, 10)
}
