% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-sampling-functions.R
\name{sample.sevt}
\alias{sample.sevt}
\title{Sample from a staged event tree}
\usage{
sample.sevt(n = 1, object)
}
\arguments{
\item{n}{number of observations}

\item{object}{the staged event tree object}
}
\value{
A data frame containing a sample of size \code{n}
}
\description{
Sample from a staged event tree
}
\details{
TO WRITE
}
