% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{set_stage}
\alias{set_stage}
\title{Set stage to path}
\usage{
set_stage(sevt, path, stage)
}
\arguments{
\item{sevt}{Staged event tree}

\item{path}{Vector of the path}

\item{stage}{stage to be assigned}
}
\description{
Set the given stage to the path for the stage event tree
}
\keyword{internal}
