% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-base-model-function.R
\name{sevt class}
\alias{sevt class}
\title{Staged event tree (sevt) class}
\description{
Structure and usage of S3 class \code{sevt}
used to store a staged event tree.
}
\details{
A staged event tree object is a list with components:
\itemize{
\item tree (required): A named list where for each variable,
the levels of such variable are listed.
The order of the variable is the
order of the event tree.
\item stages (required): A named list where each component
stores the stages for the given variable.
\item ctables: The contingency tables of the data distributed
along the tree structure.
\item lambda: The smoothing parameter used to compute probabilities.
\item prob: The conditional probability tables for every
variable and stage.
\item ll: The log-likelihood of the \code{estimated} model.
}
The tree structure is never defined explicitly, but instead
it is defined by the list \code{tree} containing the order
of the variables and the names of their levels. This is
sufficient to define a complete symmetric tree where an
internal node at a depth related to a variable \code{v}
has a number of children equal to the cardinality of
the levels of \code{v}.
The stages informations is instead stored as a list of
vectors, where each vector is indexed as the internal nodes
of the tree at a given depth.

To define a staged tree from data (data frame or table) the
user can call either \code{\link{full}} or \code{\link{indep}}
which both construct the staged tree object, attach the data in
\code{ctables} and compute probabilities. After, one of the
available model selection algorithm can be used, see for example
\code{\link{hc}}, \code{\link{bhc}} or
\code{\link{stages_hclust}}.
If, mainly for development, only the staged tree structure is needed (without data or
probabilities) the basic \code{\link{staged_ev_tree}} constructor can
be used.
}
