% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{sevt.fit}
\alias{sevt.fit}
\title{Fit a staged event tree}
\usage{
sevt.fit(object, data = NULL, lambda = object$lambda)
}
\arguments{
\item{object}{a staged event tree.}

\item{data}{data.frame or contingency table with observations of
the variables in \code{object}.}

\item{lambda}{smoothing parameter.}
}
\value{
a fitted staged event tree, that is an object of class \code{sevt}
with \code{ctables} and \code{prob} arguments.
}
\description{
Estimate transition probabilities in a staged event tree from data.
Probabilities are estimated with the relative frequencies plus,
eventually,  an additive (Laplace) smoothing.
}
\details{
The log-likelihood of the model will be recomputed and
stored in the returned object.
}
\examples{

#########
model <- staged_ev_tree(list(
  X = c("good", "bad"),
  Y = c("high", "low")
))
D <- data.frame(
  X = c("good", "good", "bad"),
  Y = c("high", "low", "low")
)
model.fit <- sevt.fit(model, data = D, lambda = 1)
}
