% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-base-model-function.R
\name{sevt}
\alias{sevt}
\alias{sevt.table}
\alias{sevt.data.frame}
\alias{sevt.list}
\title{Build a staged event tree}
\usage{
sevt(x, full = FALSE, order = NULL)

\method{sevt}{table}(x, full = FALSE, order = names(dimnames(x)))

\method{sevt}{data.frame}(x, full = FALSE, order = colnames(x))

\method{sevt}{list}(x, full = FALSE, order = names(x))
}
\arguments{
\item{x}{a list, a data frame or table object.}

\item{full}{logical, if TRUE the full model is created
otherwise the independence model.}

\item{order}{character vector,
order of the variables to build the
tree, by default the order of the variables
in \code{x}.}
}
\value{
A staged event tree object, element of \code{sevt}
class.
}
\description{
Basic constructor for staged event tree class (\code{sevt}).
}
\examples{

######### from table
model.titanic <- sevt(Titanic, full = TRUE) 

######### from data frame
DD <- generate_random_dataset(n = 4, 1000)
model.indep <- sevt(DD)
model.full <- sevt(DD, full = TRUE)

######### from list
model <- sevt(list(
  X = c("good", "bad"),
  Y = c("high", "low")
))
}
