% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_info.R
\name{sevt_df}
\alias{sevt_df}
\title{Number of parameters of a staged event tree}
\usage{
sevt_df(x)
}
\arguments{
\item{x}{An object of class \code{sevt}.}
}
\value{
integer, degrees of freedom of the staged event tree.
}
\description{
Return the number of parameters of the model.
}
