% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_info.R
\name{sevt_varnames}
\alias{sevt_varnames}
\title{Variable names}
\usage{
sevt_varnames(object)
}
\arguments{
\item{object}{an object of class \code{sevt}.}
}
\value{
A character vector.
}
\description{
Utility returning variable-names in a staged event tree
model.
}
\keyword{internal}
