% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{split_stage_random}
\alias{split_stage_random}
\title{split randomly a stage}
\usage{
split_stage_random(object, stratum, stage, p = 0.5)
}
\arguments{
\item{object}{a staged event tree object}

\item{stratum}{the variable where to split the stage}

\item{stage}{the name of the stage}

\item{p}{probability}
}
\value{
a staged event tree object
}
\description{
Randomly assign some of the path to a new stage
}
