% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{staged_ev_tree.data.frame}
\alias{staged_ev_tree.data.frame}
\title{Staged (stratified) event tree}
\usage{
\method{staged_ev_tree}{data.frame}(x, order = colnames(x),
  full = FALSE, fit = TRUE, ...)
}
\arguments{
\item{x}{data.frame with the observation}

\item{order}{vector of order, default to the order of the columns of \code{x}}

\item{full}{logical, if the full model should be built instead}

\item{fit}{logical}

\item{...}{additional parameters to be passed to other methods (mainly 
\link{fit.staged_ev_tree})}
}
\value{
A staged event tree object
}
\description{
Builds the staged event tree from a data set of categorical variables,
the order can be specified.
}
\details{
The staged event trees
}
\examples{
DD <- generate_random_dataset(n = 5, 1000)
sevt <- staged_ev_tree(DD, fit = TRUE, full = FALSE)
}
