% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-base-model-function.R
\name{staged_ev_tree.list}
\alias{staged_ev_tree.list}
\title{Build a staged event tree}
\usage{
\method{staged_ev_tree}{list}(x, full = FALSE)
}
\description{
Build a staged event tree
}
\examples{

######### from list
model <- staged_ev_tree(list(
  X = c("good", "bad"),
  Y = c("high", "low")
))
}
