% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{staged_ev_tree}
\alias{staged_ev_tree}
\title{Staged (stratified) event tree}
\usage{
staged_ev_tree(x, order = NULL, fit = FALSE, ...)
}
\arguments{
\item{x}{data.frame or list}

\item{order}{Vector of variables names, the order to build the event tree}

\item{fit}{If `TRUE` the conditional probability will be estimated from `data`}
}
\value{
A staged event tree object (see Details)
}
\description{
Builds the staged event tree for a set of categorical variables,
the order can be specified.
}
\details{
A staged event tree object consist
}
