% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{staged_ev_tree}
\alias{staged_ev_tree}
\title{Staged (stratified) event tree}
\usage{
staged_ev_tree(x, ...)
}
\arguments{
\item{x}{data.frame, list or stratified event tree object}

\item{...}{additional parameters to be passed
to the appropiate method, see \link{staged_ev_tree.data.frame}}
}
\value{
A staged event tree object (see Details)
}
\description{
Builds the staged event tree for a set of categorical variables,
the order can be specified.
}
\details{
A staged event tree object is a list with components:
\itemize{
         \item tree: A named list where for each variable, the levels of
                 such variable are listed. The order of the variable is the
                 order of the event tree.
         \item stages: A named list where each component stores the stages
                   for the given variable, stages are represented by numbers
                   that do not need to be consecutive.
         \item paths: A named list where for each variable we list in a
         data.frame object all the possible paths up to that point and the
         assigned stages.
         \item prob: The conditional probability tables for every variable
         and every stage (present only if the staged event tree has been
         fitted)
         }
}
