% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{stageinfo.sevt}
\alias{stageinfo.sevt}
\title{Print stage(s) info}
\usage{
stageinfo.sevt(object, var, stage = NULL)
}
\arguments{
\item{object}{a staged event tree object}

\item{var}{name of one variable}

\item{stage}{name(s) of stage for variable \code{var}}
}
\value{
a print of most important informations associated to stages, for a given variable.
If \code{stage} is specified, informations of only that stage are printed, otherwise
the informations of all stages related to the specified variable are printed.
}
\description{
Print stage(s) info
}
\examples{
model <- full(PhDArticles, fit = TRUE)
model <- fbhc.sevt(model)
stageinfo.sevt(model, var = "Mentor")
stageinfo.sevt(model, var = "Mentor", stage = 24)
}
