% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{stages.sevt}
\alias{stages.sevt}
\title{Print stages info}
\usage{
stages.sevt(object, var = NULL)
}
\arguments{
\item{object}{a staged event tree object}

\item{var}{name of one variable}
}
\description{
Print stages info
}
