% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-model-selection.R
\name{stages_hc}
\alias{stages_hc}
\title{Hill-Climb Score optimization}
\usage{
stages_hc(
  object,
  score = function(x) {     return(-BIC(x)) },
  max_iter = Inf,
  scope = NULL,
  ignore = object$name_unobserved,
  trace = 0
)
}
\arguments{
\item{object}{a staged event tree object.}

\item{score}{a function that score staged event tree objects.}

\item{max_iter}{the maximum number of iterations per variable.}

\item{scope}{names of variables that should be considered for the optimization}

\item{ignore}{vector of stages which will be ignored and left untouched,
by default the name of the unobserved stages stored in
\code{object$name_unobserved}.}

\item{trace}{integer, if positive information on the progress is
printed to console}
}
\value{
The final staged event tree obtained.
}
\description{
Hill-climbing search of staged event trees with
iterative moving of nodes between stages.
}
\details{
For each variable node-moves that best increase the
score are performed until no increase is possible.
A node-move is either changing the stage
associate to a node or move the node to a new stage.

The \code{ignore} argument can be used to specify stages that should not
be affected during the search, that is left untouched.
This is usefull for preserving structural zeroes and to speed-up
computations.
}
\examples{
model <- stages_hc(full(PhDArticles[, 1:3], lambda = 1))
summary(model)

## preserve zero stages
start <- indep(PhDArticles[,1:5], join_unobserved = TRUE)
model <- stages_hc(start)
}
