% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-model-selection.R
\name{stages_hclust}
\alias{stages_hclust}
\title{Learn a staged tree with hierarchical clustering}
\usage{
stages_hclust(
  object,
  distance = "totvar",
  k = length(object$tree[[1]]),
  method = "complete",
  ignore = object$name_unobserved,
  limit = length(object$tree),
  scope = NULL
)
}
\arguments{
\item{object}{a staged event tree object.}

\item{distance}{string, the distance measure to be used, either
a possible \code{method} for \code{\link{dist}} or
one of the following: \code{"totvar", "hellinger"}.}

\item{k}{integer or (named) vector: number of clusters, that is stages per variable.
Values will be recycled if needed.}

\item{method}{the agglomeration method to be used in \code{\link{hclust}}.}

\item{ignore}{vector of stages which will be ignored and left untouched,
by default the name of the unobserved stages stored in
\code{object$name_unobserved}.}

\item{limit}{the maximum number of variables to consider.}

\item{scope}{names of the variables to consider.}
}
\value{
A staged event tree object.
}
\description{
Build a stage event tree with \code{k} stages for each variable by
clustering stage probabilities with hierarchical clustering.
}
\details{
\code{hclust_sevt} performs hierarchical clustering
of the initial stage probabilities in \code{object}
and it aggregates them into the specified number
of stages (\code{k}).
A different number of stages for the different variables
in the model can be specified by supplying a (named) vector
via the argument \code{k}.
}
\examples{
data("Titanic")
model <- stages_hclust(full(Titanic, join_unobserved = TRUE, lambda = 1), k = 2)
summary(model)
}
