% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{stndnaming.sevt}
\alias{stndnaming.sevt}
\title{Standard renaming of stages}
\usage{
stndnaming.sevt(object, rep = FALSE)
}
\arguments{
\item{object}{a staged event tree object}

\item{rep}{logical, if stages name can be repeated for different
variables}
}
\value{
a staged event tree object with stages named with
consecutive integers.
}
\description{
Standard renaming of stages
}
