% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{stndnaming.sevt}
\alias{stndnaming.sevt}
\title{Standard renaming of stages}
\usage{
stndnaming.sevt(object, uniq = FALSE, prefix = FALSE, ignore = NULL)
}
\arguments{
\item{object}{a staged event tree object}

\item{uniq}{logical, if stage numbers shuold be unique over all tree}

\item{prefix}{logical, if stage names should be prefixed with variable name}

\item{ignore}{stages name which should be left untouched}
}
\value{
a staged event tree object with stages named with
consecutive integers.
}
\description{
Standard renaming of stages
}
\examples{
DD <- generate_xor_dataset(4, 100)
model <- staged_ev_tree(DD, full = TRUE)
model <- fbhc.sevt(model)
model$stages
model1 <- stndnaming.sevt(model)
model1$stages
}
