% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{stndnaming.sevt}
\alias{stndnaming.sevt}
\title{Standard renaming of stages}
\usage{
stndnaming.sevt(object, rep = FALSE)
}
\description{
@param object a staged event tree object
@param rep logical, if stages name can be repeated for different
variables
}
