% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-base-model-function.R
\name{strt_ev_tree.sevt}
\alias{strt_ev_tree.sevt}
\title{Stratified event tree from a staged event tree}
\usage{
\method{strt_ev_tree}{sevt}(x, ...)
}
\arguments{
\item{x}{a staged event tree object}

\item{...}{additional parameters}
}
\value{
A stratified event tree object
}
\description{
Stratified event tree from a staged event tree
}
\details{
This function build a stratified event tree object
from a staged event tree.
}
