% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{subtree}
\alias{subtree}
\title{Extract subtree}
\usage{
subtree(object, path)
}
\arguments{
\item{object}{a staged event tree.}

\item{path, }{the path after which extract the subtree.}
}
\value{
A staged event tree object corresponding to the subtree.
}
\description{
Extract subtree
}
\details{
Returns the subtree of the staged event tree, starting from
\code{path}.
}
\examples{
DD <- generate_random_dataset(4, 100)
model <- staged_ev_tree(DD, full = TRUE)
plot(model)
model1 <- subtree(model, path = c("-1", "1"))
plot(model1)
}
