% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-util-functions.R
\name{tree_idx}
\alias{tree_idx}
\title{return path index}
\usage{
tree_idx(path, tree, complete = FALSE)
}
\arguments{
\item{path}{a path from root in the tree.}

\item{tree}{a symmetric tree given as a list of levels.}

\item{complete}{logical, if \code{TRUE} the complete indexing
is returned.}
}
\value{
an integer, the index of the node corresponding to \code{path}
}
\description{
return path index
}
\details{
Compute the integer index of the node associated with the
given path in a symmetric tree defined by \code{tree}.
}
\keyword{internal}
