% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-base-model-function.R
\name{variable.names.sevt}
\alias{variable.names.sevt}
\title{Variable names}
\usage{
\method{variable.names}{sevt}(object, ...)
}
\arguments{
\item{object}{a staged event tree object}

\item{...}{further arguments passed to or from other methods}
}
\value{
A character vector
}
\description{
Utility returning variable-names in a staged event tree
model.
}
\examples{

mod <- full(PhDArticles)
variable.names(mod)
}
