% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-util-functions.R
\name{which_class}
\alias{which_class}
\title{find maximum value}
\usage{
which_class(x, levels)
}
\arguments{
\item{x}{numerical, the log-probabilities}

\item{character}{the levels to be returned same lenght as x}
}
\value{
factor
}
\description{
find maximum value
}
\keyword{internal}
